### 2024-02-13 #157 @isabelizimm
---
---

As a Floridian, I do appreciate your tutorial locations 🐊

A few quick fixes:

For the `CODE_OF_CONDUCT` file, it is optimal to have it at the root of the repository. Right now, it looks like yours is in `docs/source/Code of Conduct.rst`. I'd recommend moving that file, since that is the typical place people look for a CoC. Also, if it is in the root, it will show up as a "tab" next to your README, sort of how the MIT License is shown here 🎉

Second, pending some sort of tool that requires it, you shouldn't need a separate `[metadata]` [section in your `pyproject.toml`](https://github.com/USEPA/harmonize-wq/blob/732c69ee71cba83ad63dc11ea268a246ee9a11a0/pyproject.toml#L8).
