+++ pyproject.toml
@@ -5,10 +5,6 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools]
 packages = ["harmonize_wq"]
 
-[metadata]
-name = "harmonize_wq"
-version = "0.4.0"
-
 [project]
 name = "harmonize_wq"
 version = "0.4.0"
