+++ .github/workflows/documentation_test.yml
@@ -6,7 +6,7 @@ on:
       [main]
   push:
     branches:
-      [main, dev]
+      [main, dev, pyOpenSci-review]
 
 # Allows you to run this workflow manually from the Actions tab
 #workflow_dispatch:
