### 2024-02-14 #157 @jbousquin
Thanks @isabelizimm - made those suggested changes on [pyOpenSci-review](https://github.com/USEPA/harmonize-wq/tree/pyOpenSci-review) branch. Let me know if there is anything else while we wait.
