### 2024-03-16 #157 @Batalex
:wave: Hi @rcaneill and @jacqui-123! Thank you for volunteering to review for pyOpenSci!

Please don't hesitate to introduce yourselves. @jbousquin, I am pleased to announce that we found our A-team to proceed with the review.

- [ ] @rcaneill survey completed.
- [ ] @jacqui-123 survey completed.

1. Here is the **[reviewers guide](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html)**. This guide contains all the steps and information needed to complete your review.
2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

Please get in touch with any questions or concerns! Your review is due: April 8th
---

Reviewers: @rcaneill, @jacqui-123
Due date: 2024/04/08
