### 2024-03-31 #157 @Batalex
Please find below a list of comments.
I tried to rank them so that you can prioritize your work.
I'll complete this list as I revisit the package.

## Praises
- praise (general): The code and the docs are extra clean.
- praise (general): Whenever I see `pint`, I'm happy!

## Typos
- typo (readme.md): l7 on package name
- typo (readme.md, contributing.rst): double spaces

## Nitpicks
- nitpick (general): I recommend adding a new line at each full stop in a markdown or rst paragraph. This way, we keep the lines short in git (easier to spot diffs in PR, easier to pinpoint a line with an issue). No worries, a single new line is not rendered.
- nitpick (domain.py): there is no need for a raw string for `TADA_DATA_URL`

## Discussions
- discussion (convert.py): About the TODO - both points of view (regrouping constants in a single place, or having them defined near their place of use to avoid jumping around the code base) are valid. I am usually in favor of the former.

## Suggestions
- suggestion (domain.py): In `harmonize_TADA_dict`, we could use a groupby operation to avoid looping through the dataframe using python. TOCHECK
- suggestion (domain.py): We could replace the following pattern `for x in list(set(pandas_series))` by using the `.unique` method
- suggestion (domain.py, basic.py): `out_col_lookup` does not need to be a function. Same for all other functions returning a dict. If we make those simple module-level dicts, we can still list the sources in the module docstring.
- suggestion (convert.py): We could add "references" sections in the docstrings so that the sources are present in the website and not only in the source code.
- suggestion (basis.py, general): By using pandas' methods, we could streamline a little some operations. The choice is ultimately yours; I prefer using existing methods over rolling my own implementations, even if that means that other folks need to go to the documentation website to understand what is going on.
For instance, here is my proposition for `set_basis`
```py
def set_basis(df, mask, basis, basis_col):
    return df.assign(**{basis_col: np.where(mask, basis, np.nan)})
```
I find this implementation easier to read (but I understand that this is debatable), but it is also more efficient. I have noticed that you use this pattern quite a few time throughout the code base, so I figured this might interests you.

## Todos
- todo (pyproject.toml): We should remove the `metadata` section.
- todo (`__init__.py`): `importlib.metadata` was added in python 3.8, which is the minimal version supported by the package according to its `pyproject.toml`. The try .. except block should not be needed, even more so considering that `importlib_metadata` is not listed in the project requirements.
- todo (basis.py): We could regroup the conditions branches in `update_result_basis`
- todo (contributing.rst): To lower the cost of entry for potention contributors, let's make sure that we provide all the information they need. Consider adding a section describing how to setup their development environment (e.g. installing the test and docs dependencies).

## Issues
- issue (general): code quality (see below)
- issue (domain.py): `requests` should be listed in the project's dependencies. The rationale is as follows: we should not import in our code any transitive dependency, because we have no guarantee that the primary dependency will not drop the former in a future update. As far as we know, `dataretrieval` could replace `requests` by `httpx` without notice in a patch release, which would break new `harmonize-wq`installations. The same can be said about `pandas`, though I agree it is unlikely that `geopandas` will change its backend dataframe lib.
- issue (domain.py): We should specify what kind of exception we are expecting in `re_case`. Making a try except block too wide can lead to hard-to-debug issues.
- issue (general): It seems that there are circular dependencies: harmonize -> visualize -> wrangle -> harmonize or clean -> wrangle -> clean as well. They do not raise an exception for now, but they will if any imported object is used at the module level. I strongly advise that we rework the project structure so that the files get imported in an acyclic fashion. It is also way easier to get familiar with the code base as a new contributor if the structure is predictable and linear.
-

## General recommendations

Code quality is important in a public package.
It is obvious that a great amount of care went in making `harmonize_wq`, but what I mean by code quality is having tools enforcing conventions across the code base.
Such conventions usually cover code format, and catching simple anti patterns.

To do so, I would advise you to use both a linter and a formatter.
I usually recommend:
- [black](https://black.readthedocs.io/en/stable/index.html) for formatting the code
- [ruff](https://docs.astral.sh/ruff/) to validate that the code follows good practices, and do quick fixes.

This is up to debate of course, some people might prefer one tool over another, but the point is that a project using such tools:
- is more welcoming to external contributors
- needs less time dedicated to low-value maintainance.

If you are ok with everything I said so far, I'd be happy to propose a PR to help you setup everything.
