### 2024-04-01 4d79ff5
Update README.md
Edit (general): new line at each full stop in a markdown paragraph.
