### 2024-04-01 8e9efd8
Update contributing.rst
fix typo (contributing.rst): double spaces
