### 2024-04-01 f86bc05
Update contributing.rst
Edit (general): adding a new line at each full stop in a rst paragraph.
