### 2024-04-01 8ea1eaa
wet_dry_drop() has become outside the normal workflow, what was being done in the try/except was a bit proactive and gets problematic with keeping the functions in a linear workflow.
