### 2024-04-03 #157 @Batalex
@jbousquin Based on a quick look through the PR, yes that's exactly what I had in mind
