### 2024-04-03 daba3b1
chore: Inital setup for automated code quality
