+++ harmonize_wq/domains.py
@@ -91,6 +91,7 @@ def get_domain_dict(table, cols=None):
     Return dictionary for domain from WQP table (e.g., 'ResultSampleFraction'),
     The default keys ('Name') are shown as values ('Description') are long:
     
+    >>> from harmonize_wq import domains
     >>> domains.get_domain_dict('ResultSampleFraction').keys() # doctest: +NORMALIZE_WHITESPACE
     dict_keys(['Acid Soluble', 'Bed Sediment', 'Bedload', 'Bioavailable', 'Comb Available',
                'Dissolved', 'Extractable', 'Extractable, CaCO3-bound', 'Extractable, exchangeable',
