### 2024-04-10 81448a9
Py open sci review (#56)
* Update README.md

Fix typo (readme.md): l7 on package name

* Update README.md

Fix typo (readme.md): double spaces

* Update README.md

Edit (general): new line at each full stop in a markdown paragraph.

* Update contributing.rst

fix typo (contributing.rst): double spaces

* Update contributing.rst

Edit (general): adding a new line at each full stop in a rst paragraph.

* Update example workflow.rst

Edit (general): add a new line at each full stop in a markdown or rst paragraph.

* Update index.rst

Edit (general): add a new line at each full stop in a markdown 
[…]
