### 2024-04-11 47b244a
Add development environment setup to contributing.rst:
