### 2024-04-11 1269896
Add dependency dependencies that are used directly
