+++ requirements.txt
@@ -4,3 +4,5 @@ pint>=0.18          # May work on earlier versions but this is what it was teste
 dataretrieval>=1.*, <1.0.5  # 1.0.5 will break datetime handling, <1.0 doesn't have required functions
 descartes>=1.1.0           # May be required for mapping in some jupyter notebooks
 mapclassify>=2.4.0  # May be required for mapping in some jupyter notebooks
+requests
+numpy
