### 2024-04-11 75cb733
remove try/except
