+++ harmonize_wq/__init__.py
@@ -1,9 +1,5 @@
 from harmonize_wq import harmonize
-
-try:
-  from importlib.metadata import version, PackageNotFoundError
-except ImportError:
-  from importlib_metadata import version, PackageNotFoundError
+from importlib.metadata import version, PackageNotFoundError
 
 try:
     __version__ = version('harmonize_wq')
