### 2024-04-11 01cae6e
specify exception expected in re_case.
