### 2024-04-11 d8125c4
suggestion: replace x in list(set(pandas_series)) pattern w/ .unique method
