### 2024-04-11 3544328
suggestion: replace one dict functions w/ module-level dicts (domains.out_col_lookup)
