### 2024-04-11 20922e1
suggestion: replace one dict functions w/ module-level dicts (xy_datum, stations_rename, accepted_methods)
