+++ harmonize_wq/domains.py
@@ -503,452 +503,461 @@ def characteristic_cols(category=None):
     return col_list
 
 
-def xy_datum():
-    """Get dictionary of expected horizontal datums.
+"""Get dictionary of expected horizontal datums.
 
-    The structure has {key as expected string: value as {"Description": string
-    and "EPSG": integer (4-digit code)}.
+The structure has {key as expected string: value as {"Description": string
+and "EPSG": integer (4-digit code)}.
 
-    Notes
-    -----
-    source WQP: HorizontalCoordinateReferenceSystemDatum_CSV.zip
-    
-    Anything not in dict will be NaN, and non-integer EPSG will be missing:
-    "OTHER": {"Description": 'Other', "EPSG": nan},
-    "UNKWN": {"Description": 'Unknown', "EPSG": nan}
+Notes
+-----
+source WQP: HorizontalCoordinateReferenceSystemDatum_CSV.zip
 
-    Returns
-    -------
-    dict
-        Dictionary where exhaustive:
-            {HorizontalCoordinateReferenceSystemDatumName: {Description:str,
-            EPSG:int}}
+Anything not in dict will be NaN, and non-integer EPSG will be missing:
+"OTHER": {"Description": 'Other', "EPSG": nan},
+"UNKWN": {"Description": 'Unknown', "EPSG": nan}
 
-    Examples
-    --------
-    Running the function returns the full dictionary with {abbreviation:
-    {'Description':values, 'EPSG':values}}. The abbreviation key can be used to
-    get the EPSG code:
-        
-    >>> domains.xy_datum()['NAD83']
-    {'Description': 'North American Datum 1983', 'EPSG': 4269}
-    >>> domains.xy_datum()['NAD83']['EPSG']
-    4269
-    """
-    return {"NAD27": {"Description": 'North American Datum 1927',
-                      "EPSG": 4267},
-            "NAD83": {"Description": 'North American Datum 1983',
-                      "EPSG": 4269},
-            "AMSMA": {"Description": 'American Samoa Datum',
-                      "EPSG": 4169},
-            "ASTRO": {"Description": 'Midway Astro 1961',
-                      "EPSG": 4727},
-            "GUAM": {"Description": 'Guam 1963',
-                     "EPSG": 4675},
-            "JHNSN": {"Description": 'Johnson Island 1961',
-                      "EPSG": 4725},
-            "OLDHI": {"Description": 'Old Hawaiian Datum',
-                      "EPSG": 4135},
-            "PR": {"Description": 'Puerto Rico Datum',
-                   "EPSG": 6139},
-            "SGEOR": {"Description": 'St. George Island Datum',
-                      "EPSG": 4138},
-            "SLAWR": {"Description": 'St. Lawrence Island Datum',
-                      "EPSG": 4136},
-            "SPAUL": {"Description": 'St. Paul Island Datum',
-                      "EPSG": 4137},
-            "WAKE": {"Description": 'Wake-Eniwetok 1960',
-                     "EPSG": 6732},
-            "WGS72": {"Description": 'World Geodetic System 1972',
-                      "EPSG": 6322},
-            "WGS84": {"Description": 'World Geodetic System 1984',
-                      "EPSG": 4326},
-            "HARN": {"Description": 'High Accuracy Reference Network for NAD83',
-                     "EPSG": 4152},
-            }
+Returns
+-------
+dict
+    Dictionary where exhaustive:
+        {HorizontalCoordinateReferenceSystemDatumName: {Description:str,
+        EPSG:int}}
 
+Examples
+--------
[…]
