### 2024-04-11 7962e1c
Updated implementations of the dicts vs functions.
