### 2024-04-11 b60ecf0
assign w/ np.where() for masked replace
