### 2024-04-11 3729baa
replace x in list(set(pandas_series)) pattern w/ .unique method
