### 2024-04-11 7751b1b
replace one dict functions w/ module-level dicts (unit_basis_dict(out_col) -> unit_basis_dict[out_col], basis_conversion, stp_dict).
