+++ harmonize_wq/basis.py
@@ -1,98 +1,98 @@
 # -*- coding: utf-8 -*-
 """Functions to process characteristic basis or return basis dictionary."""
+
 import numpy
 from warnings import warn
 from numpy import nan
 from harmonize_wq.clean import add_qa_flag
 
-def unit_basis_dict(out_col):
-    """Characteristic specific basis dictionary to define basis from units.
 
-    The out_col is often derived from :attr:`WQCharData.char_val`. The desired
-    basis can be used as a key to subset result.
+"""Characteristic specific basis dictionary to define basis from units.
 
-    Parameters
-    ----------
-    out_col : str
-        Column name where results are written.
+The out_col is often derived from :attr:`WQCharData.char_val`. The desired
+basis can be used as a key to subset result.
 
-    Returns
-    -------
-     dict
-         Dictionary with logic for determining basis from units string and
-         standard :mod:`pint` units to replace those with.
-         The structure is {Basis: {standard units: [unit strings with basis]}}.
+Parameters
+----------
+out_col : str
+    Column name where results are written.
 
-    Examples
-    --------
-    Get dictionary for Phosphorus and subset for 'as P':
+Returns
+-------
+ dict
+     Dictionary with logic for determining basis from units string and
+     standard :mod:`pint` units to replace those with.
+     The structure is {Basis: {standard units: [unit strings with basis]}}.
+
+Examples
+--------
+Get dictionary for Phosphorus and subset for 'as P':
+
+>>> from harmonize_wq import basis
+>>> basis.unit_basis_dict['Phosphorus']['as P']
+{'mg/l': ['mg/l as P', 'mg/l P'], 'mg/kg': ['mg/kg as P', 'mg/kg P']}
+"""
+unit_basis_dict = {
+    "Phosphorus": {
+        "as P": {"mg/l": ["mg/l as P", "mg/l P"], "mg/kg": ["mg/kg as P", "mg/kg P"]},
+        "as PO4": {
+            "mg/l": ["mg/l as PO4", "mg/l PO4"],
+            "mg/kg": ["mg/kg as PO4", "mg/kg PO4"],
+        },
+    },
+    "Nitrogen": {"as N": {"mg/l": ["mg/l as N", "mg/l N"]}},
+    "Carbon": {},
+}
+
+"""Get dictionary of conversion factors to convert basis/speciation.
+
+For example, this is used to convert 'as PO4' to 'as P'.
+
+Returns
+-------
+dict
+    Dictionary with structure {basis: conversion factor}
+     
+See Also
+--------
+:func:`convert.moles_to_mass`
+
+`Best Practices for Submitting Nutrient Data to the Water Quality eXchange
+<www.epa.gov/sites/default/files/2017-06/documents/wqx_nutrient_best_practices_guide.pdf>`_
+"""
+basis_conversion = {
+    "NH3": 0.822,
+    "NH4": 0.776,
+    "NO2": 0.304,
+    "NO3": 0.225,
+    "PO4": 0.326,
+}
+
+"""Get standard temperature and pressure to define basis from units.
+
+Notes
+-----
+    This needs to be updated to include pressure or needs to be renamed.
     
-    >>> from harmonize_wq import basis
-    >>> basis.unit_basis_dict('Phosphorus')['as P']
-    {'mg/l': ['mg/l as P', 'mg/l P'], 'mg/kg': ['mg/kg as P', 'mg/kg P']}
-    """
-    dictionary = {'Phosphorus': {'as P': {'mg/l': ['mg/l as P', 'mg/l P'],
-                                          'mg/kg': ['mg/kg as P', 'mg/kg P']},
-                                 'as PO4': {'mg/l': ['mg/l as PO4',
-                                                     'mg/l PO4'],
[…]
