### 2024-04-11 6f933f5
Implemented function -> dict changes
