### 2024-04-26 3c63d92
np.where will coerce x & y to the array dtype. A problem here as it converts np.nan to 'nan'. Reverted it back to the way it was before.
