+++ harmonize_wq/basis.py
@@ -3,7 +3,6 @@
 
 import numpy
 from warnings import warn
-from numpy import nan
 from harmonize_wq.clean import add_qa_flag
 
 
@@ -355,7 +354,7 @@ def set_basis(df_in, mask, basis, basis_col='Speciation'):
     """
     df_out = df_in.copy()
     if basis_col not in df_out.columns:
-        # Add Basis column w/ nan if it didn't exist
-        return df_out.assign(**{basis_col: numpy.where(mask, basis, nan)})
+        df_out[basis_col] = numpy.nan
     # Otherwise don't mess with existing values that are not part of mask
-    return df_out.assign(**{basis_col: numpy.where(mask, basis, df_out[basis_col])})
+    df_out.loc[mask, basis_col] = basis
+    return df_out
