### 2024-04-27 #157 @Batalex
Hey @jbousquin,
I just want to give you a brief update. @rcaneill privately reached out to me, and needs some more time to proceed with the review due to personal reasons.
Meanwhile, you can proceed with the two reviews you have here so that we avoid staling this issue for @Jacqui-123.
Does this arrangement work for you?
