### 2024-04-30 55cb3cd
Update basis.py
Revised dict doc-strings
