### 2024-05-10 #157 @rcaneill
Hi @Batalex and @jbousquin, I finished my review (cf https://github.com/pyOpenSci/software-submission/issues/157#issuecomment-2015045269)
I have 0 knowledge about the water quality field, but I found the doc quite clear :)
