### 2024-05-31 #157 @Batalex
Hey @jbousquin,
I noticed that this review has been quite stale lately, and so has `harmonize_wq`'s codebase.

Would you mind giving us a rough rundown on how and when you plan to address the reviews?
My goal here is to set the proper expectations for everyone and manage our reviewers' time effectively.
