### 2024-07-01 #157 @Batalex
Hello @jbousquin,

 I sent you a reminder a month ago about the review going stale, and I have not seen any public activity on the repository ever since.

As I said before, the deadlines in our process are more like guidelines as to when we expect things to move forward. However, as the editor for this submission, I have a responsibility to the volunteers who gave their personal time to do an in-depth review of harmonize-wq, and even submitted PRs. It is okay to be late, but I expect you to be transparent and committed to moving forward with the review.
Per our [review policy](https://www.pyopensci.org/software-peer-review/how-to/editors-guide.html#putting-a-review-on-hold-handling-non-responsive-authors), I am putting this submission on hold, and will close it one month from now on if I do not see any change.
Thank you for your understanding.
