### 2024-07-02 7446dfc
Update requirements.txt
Be more specific about lowest version
