### 2024-07-02 bfbe941
Update requirements.txt
pin older version (32 vs 64 changes for dtype)
