+++ requirements.txt
@@ -1,3 +1,4 @@
+numpy<2.0
 pandas<2.0
 geopandas>=0.10.2, <0.13           # Has additional dependencies (pandas, numpy, shapely, fiona, etc.), tested on 0.10.2
 pint>=0.18          # May work on earlier versions but this is what it was tested on
@@ -5,4 +6,3 @@ dataretrieval>=1.0, <1.0.5  # 1.0.5 will break datetime handling, <1.0 doesn't h
 descartes>=1.1.0           # May be required for mapping in some jupyter notebooks
 mapclassify>=2.4.0  # May be required for mapping in some jupyter notebooks
 requests
-numpy
