### 2024-07-02 USEPA/harmonize-wq#76 @jbousquin
AttributeError: 'float' object has no attribute 'upper'
From reviewer: attribute errors for harmonize_all(df, errors='ignore'): AttributeError: 'float' object has no attribute 'upper' (these attribute errors happened a few times in the other demos, too.)
