### 2024-07-02 ceef5c3
Minor edits to text
