### 2024-07-02 f345018
Install from deve version not pinned release (one less thing to update)
