### 2024-07-02 c3f4944
Resolve first error when knit locally
