### 2024-07-02 6a6e832
Comment out some of the option 2 code so it doesn't run when knit-ing. Open to better options of how to format it as code but not always run it..
