### 2024-07-02 c72f069
First attempt to render the rmd in R (i.e. run the python code in R vignette)
