+++ .github/workflows/test_r.yaml
@@ -38,7 +38,9 @@ jobs:
         uses: r-lib/actions/setup-r@v2
         with:
           r-version: 'release'
-
+      
+      - uses: r-lib/actions/setup-renv@v2
+      
       - uses: r-lib/actions/setup-pandoc@v2
       
       - name: Update pip and install testing pkgs
@@ -60,9 +62,14 @@ jobs:
       
       - name: Run pip env using R reticulate
         run: |
-          install.packages("reticulate")
+          install.packages(c("knitr", "rmarkdown", "reticulate"))'
           reticulate::import("harmonize_wq")
         shell: Rscript {0}
+        
+      - name: Render-rmd
+        run: |
+          echo "Rendering the page..."
+          Rscript -e 'rmarkdown::render(input = "demos/harmonize_Pensacola.Rmd")'
         
       - name: Run pytest
         run: pytest -v harmonize_wq
