### 2024-07-02 15cd379
Update test_r.yaml
Try without renv, fixed typo.
