### 2024-07-02 de53490
Update test_r.yaml
Switch from pip to conda (required of reticulate)
