+++ .github/workflows/test_r.yaml
@@ -29,10 +29,13 @@ jobs:
     steps:
       - uses: actions/checkout@v3
          
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+      - name: Install/set-up conda
+        uses: conda-incubator/setup-miniconda@v3
         with:
+          auto-update-conda: true
           python-version: ${{ matrix.python-version }}
+          auto-activate-base: true
+          activate-environment: ""
 
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
@@ -42,25 +45,18 @@ jobs:
       #- uses: r-lib/actions/setup-renv@v2
       
       - uses: r-lib/actions/setup-pandoc@v2
-      
-      - name: Update pip and install testing pkgs
+            
+      - name: Install testing pkgs
         run: |
-          python -VV
-          python -m pip install --upgrade pip
-          pip install pytest
+          conda install pytest
+          conda install geopandas>=0.10.2, <0.13 
+          conda install numpy<2.0
+          conda install pandas<2.0
+          conda install pint>=0.18
+          conda install dataretrieval>=1.0, <1.0.5
+          conda install requests
       
-      # fiona doesn't have wheels for windows
-      - if: matrix.os == 'windows-latest'
-        run: |
-          pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2023.7.16/GDAL-3.7.1-cp39-cp39-win_amd64.whl
-          pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2023.7.16/Fiona-1.9.4.post1-cp39-cp39-win_amd64.whl
-      
-      - name: Install package and dependencies
-        run: |
-          python -m pip install --no-deps .
-          pip install -r requirements.txt
-      
-      - name: Run pip env using R reticulate
+      - name: Run env using R reticulate
         run: |
           install.packages(c("knitr", "rmarkdown", "reticulate"))
           reticulate::import("harmonize_wq")
