### 2024-07-02 4c6b60a
Update test_r.yaml
Try setting up env using reticulate instead of conda (since conda-incubator is currently restricted)
