### 2024-07-02 2c23f0f
Update test_r.yaml
Combine install/run so env is not lost
