### 2024-07-02 821e90b
Update test_r.yaml
update env name
