### 2024-07-02 5f91eb5
Update test_r.yaml
use miniconda_install() (note before I was setting up venv)
