### 2024-07-02 9295259
Update test_r.yaml
install dev package from repo... for tests this should use the currently checked out instead, but just trying to get it working.
