+++ .github/workflows/test_r.yaml
@@ -68,6 +68,8 @@ jobs:
           #  packages
           #)
           #use_virtualenv("wq_harmonize")
+          
+          reticulate::py_install("git+https://github.com/USEPA/harmonize-wq.git", pip = TRUE, envname = "wq_harmonize")
 
           rmarkdown::render(input = "demos/harmonize_Pensacola.Rmd")
         
