### 2024-07-02 a2d590c
Update test_r.yaml
Not sure why now it was case sensitive...
