### 2024-07-03 78dad2c
Update test_r.yaml
pytest already check tests in test.yml, no need to run it a second time here.
