+++ .github/workflows/test_r.yaml
@@ -72,6 +72,3 @@ jobs:
           reticulate::py_install("git+https://github.com/USEPA/harmonize-wq.git", pip = TRUE, envname = "wq_harmonize")
 
           rmarkdown::render(input = "demos/Harmonize_Pensacola.Rmd")
-        
-      - name: Run pytest
-        run: pytest -v harmonize_wq
