### 2024-07-03 b104c36
Update test_r.yaml
Skip unbutu, run all windows (conda)
