+++ .github/workflows/test_r.yaml
@@ -21,11 +21,8 @@ jobs:
       
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [windows-latest, macos-latest]
         python-version: ['3.8', '3.9', '3.10', '3.11']
-        include:
-          - os: windows-latest
-            python-version: "3.9"
     steps:
       - uses: actions/checkout@v3
 
