### 2024-07-03 0f16f2d
Update test_r.yaml
Trying on ubuntu again
