### 2024-07-03 d7661c0
Update test_r.yaml
Try on older ubuntu version.
