### 2024-07-03 3316530
Update test_r.yaml
bump action version up (getting warnings)
