+++ .github/workflows/test_r.yaml
@@ -24,10 +24,10 @@ jobs:
         os: [windows-latest, macos-latest, ubuntu-20.04.2]
         python-version: ['3.8', '3.9', '3.10', '3.11']
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - name: setup python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.config.python }}
 
