### 2024-07-03 522e998
Update test_r.yaml
Drop Ubuntu. Fix where python version is set from the matrix. Conda create with python version from matrix.
