+++ .github/workflows/test_r.yaml
@@ -21,15 +21,15 @@ jobs:
       
     strategy:
       matrix:
-        os: [windows-latest, macos-latest, ubuntu-20.04.2]
+        os: [windows-latest, macos-latest]
         python-version: ['3.8', '3.9', '3.10', '3.11']
     steps:
       - uses: actions/checkout@v4
 
-      - name: setup python
+      - name: setup python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.config.python }}
+          python-version: ${{ matrix.python-version }}
 
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
@@ -56,7 +56,7 @@ jobs:
               )
           
           reticulate::install_miniconda()
-          reticulate::conda_create("wq_harmonize")
+          reticulate::conda_create("wq_harmonize", python_version=${{ matrix.python-version }})
           reticulate::conda_install("wq_harmonize", packages)
 
           #path_to_venv <- virtualenv_create(
