### 2024-07-03 7ce2925
Update test_r.yaml
Try setting var from matrix using with since it seemed to cut out 3.1* for 3.10.
