### 2024-07-03 da72c56
Update test_r.yaml
Try again (hoping it won't truncate python 3.1x this time.
