### 2024-07-03 e16aad9
Update test_r.yaml
try as sting?
