### 2024-07-05 b125f65
Dynamic config pyproject.toml
Use dynamic to specify depends from requirements.txt. This should be more maintainable. and may resolve some of the install issues reveries were having. setuptools>=61.0 already required.
