### 2024-07-05 USEPA/harmonize-wq#76 @jbousquin
Possible repeat of #64, which leads me to believe dependencies aren't being resolved properly. Bringing pinned versions up to date (#26 ) may help resolve. Updated pyproject.toml to sync with requirements (b125f65631d2395dcf3c15a3b3444afdaafd7389). Sounds like it could be a numpy.nan for a text field (trying to uppercase an NAN), this was seen in domains.re_case() and is the reason for the try/except there.
