### 2024-07-05 4042da1
Update test_r.yaml
Upload artifact at end to examine rendered html
