### 2024-07-05 c786242
Update artifact test_r.yaml
Based on examples on repo
