+++ .github/workflows/test_r.yaml
@@ -73,5 +73,6 @@ jobs:
       - name: Upload artifact
         uses: actions/upload-artifact@v4
         with:
+          name: demos-artifact
           # Upload entire demos folder
-          path: '/demos'
+          path: /demos
