### 2024-07-05 0d5c07b
Update test_r.yaml
Only upload on one runner (win 3.11
