+++ .github/workflows/test_r.yaml
@@ -71,6 +71,7 @@ jobs:
           rmarkdown::render(input = "demos/Harmonize_Pensacola.Rmd")
 
       - name: Upload artifact
+        if: ${{ (matrix.os == 'windows-latest') && (matrix.python-version == 3.11) }}
         uses: actions/upload-artifact@v4
         with:
           name: demos-artifact
