### 2024-07-05 594dc10
Wordsmith Harmonize_Pensacola.Rmd
Updated text to be more accurate with how reticulate can be used
