### 2024-07-05 7e187a2
merge from main (#81)
* Update broken link

* Update Harmonize_Pensacola.Rmd

* remove full path

* Update requirements.txt

Be more specific about lowest version

* Update requirements.txt

pin older version (32 vs 64 changes for dtype)

* Update domains.py

Interesting new domain was added...

* Minor edits to text

* Fix plot

* Install from deve version not pinned release (one less thing to update)

* Resolve first error when knit locally

* Comment out some of the option 2 code so it doesn't run when knit-ing. Open to better options of how to format it as code but not always run it..

* F
[…]
