+++ demos/Harmonize_Pensacola.Rmd
@@ -1,112 +1,174 @@
 ---
 title: "harmonize-wq in R"
+format: html
+editor: visual
 author: "Justin Bousquin, Cristina Mullin, Marc Weber"
-date: '2022-08-31'
-output: rmarkdown::html_vignette
+date: "`r Sys.Date()`"
+output: 
+  rmarkdown::html_vignette:
+    toc: true
+    fig_caption: yes
+    fig_height: 8
+    fig_width: 8
 vignette: >
+  %\VignetteEncoding{UTF-8}
   %\VignetteIndexEntry{harmonize-wq in R}
-  %\usepackage[utf8]{inputenc}
   %\VignetteEngine{knitr::rmarkdown}
-editor_options: 
+editor_options:
   chunk_output_type: console
+  markdown:
+    wrap: 72
 ---
 
 ```{r setup, include = FALSE}
-# Set chunk options
+library(knitr)
+
 knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
+  echo = TRUE,
+  warning = FALSE,
+  message = FALSE
 )
 ```
 
-<br>
-
 ## Overview
 
-Standardize, clean, and wrangle Water Quality Portal data into more analytic-ready formats using the harmonize_wq package. US EPA’s Water Quality Portal (WQP) aggregates water quality, biological, and physical data provided by many organizations and has become an essential resource with tools to query and retrieval data using python or R. Given the variety of data and variety of data originators, using the data in analysis often requires data cleaning to ensure it meets the required quality standards and data wrangling to get it in a more analytic-ready format. Recognizing the definition of analysis-ready varies depending on the analysis, the harmonize_wq package is intended to be a flexible water quality specific framework to help:
+Standardize, clean, and wrangle Water Quality Portal data into more
+analytic-ready formats using the harmonize_wq package. US EPA's Water
+Quality Portal (WQP) aggregates water quality, biological, and physical
+data provided by many organizations and has become an essential resource
+with tools to query and retrieval data using python or R. Given the
+variety of data and variety of data originators, using the data in
+analysis often requires data cleaning to ensure it meets the required
+quality standards and data wrangling to get it in a more analytic-ready
+format. Recognizing the definition of analysis-ready varies depending on
+the analysis, the harmonize_wq package is intended to be a flexible
+water quality specific framework to help:
 
-* Identify differences in data units (including speciation and basis)
-* Identify differences in sampling or analytic methods
-* Resolve data errors using transparent assumptions
-* Reduce data to the columns that are most commonly needed
-* Transform data from long to wide format
+-   Identify differences in data units (including speciation and basis)
+-   Identify differences in sampling or analytic methods
+-   Resolve data errors using transparent assumptions
+-   Reduce data to the columns that are most commonly needed
+-   Transform data from long to wide format
 
-Domain experts must decide what data meets their quality standards for data comparability and any thresholds for acceptance or rejection.
-
-<br>
-
-<br>
+Domain experts must decide what data meets their quality standards for
+data comparability and any thresholds for acceptance or rejection.
 
 ## Installation & Setup
 
-#### Install the harmonize-wq package (Command Line)
+Using R/reticulate requires an installation of Python to bind to, and EPA's harmonize-wq package must be run from a python environment with the packages it depends on.
+
+For environment management, reticulate requires either conda (recommended) or virtualenv.
+
[…]
