+++ .github/workflows/test_r.yaml
@@ -21,16 +21,13 @@ jobs:
       
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [windows-latest, macos-latest]
         python-version: ['3.8', '3.9', '3.10', '3.11']
-        include:
-          - os: windows-latest
-            python-version: "3.9"
     steps:
-      - uses: actions/checkout@v3
-         
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+      - uses: actions/checkout@v4
+
+      - name: setup python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
 
@@ -38,31 +35,45 @@ jobs:
         uses: r-lib/actions/setup-r@v2
         with:
           r-version: 'release'
-
+      
+      #- uses: r-lib/actions/setup-renv@v2
+      
       - uses: r-lib/actions/setup-pandoc@v2
       
-      - name: Update pip and install testing pkgs
-        run: |
-          python -VV
-          python -m pip install --upgrade pip
-          pip install pytest
-      
-      # fiona doesn't have wheels for windows
-      - if: matrix.os == 'windows-latest'
-        run: |
-          pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2023.7.16/GDAL-3.7.1-cp39-cp39-win_amd64.whl
-          pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2023.7.16/Fiona-1.9.4.post1-cp39-cp39-win_amd64.whl
-      
-      - name: Install package and dependencies
-        run: |
-          python -m pip install --no-deps .
-          pip install -r requirements.txt
-      
-      - name: Run pip env using R reticulate
-        run: |
-          install.packages("reticulate")
-          reticulate::import("harmonize_wq")
+      - name: R depends
         shell: Rscript {0}
+        run: |
+          install.packages(c("knitr", "rmarkdown", "reticulate"))
         
-      - name: Run pytest
-        run: pytest -v harmonize_wq
+      - name: setup r-reticulate venv & render rmd
+        shell: Rscript {0}
+        run: |
+          library(reticulate)
+
+          packages = c(
+              "pytest", "numpy<2.0", "pandas<2.0", "geopandas>=0.10.2, <0.13", "pint>=0.18",
+              "dataretrieval>=1.0, <1.0.5", "requests"
+              )
+          
+          reticulate::install_miniconda()
+          reticulate::conda_create("wq_harmonize", python_version = "${{ matrix.python-version }}")
+          reticulate::conda_install("wq_harmonize", packages)
+
+          #path_to_venv <- virtualenv_create(
+          #  envname = "wq_harmonize",
+          #  python = Sys.which("python"), # placed on PATH by the setup-python action
+          #  packages
+          #)
+          #use_virtualenv("wq_harmonize")
+          
+          reticulate::py_install("git+https://github.com/USEPA/harmonize-wq.git", pip = TRUE, envname = "wq_harmonize")
+
+          rmarkdown::render(input = "demos/Harmonize_Pensacola.Rmd")
+
+      - name: Upload artifact
+        if: ${{ (matrix.os == 'windows-latest') && (matrix.python-version == 3.11) }}
+        uses: actions/upload-artifact@v4
+        with:
+          name: demos-artifact
+          # Upload entire demos folder
+          path: './demos'
