### 2024-07-05 4d1dd8c
Update contributing.rst
Adding example code for dev env. Closes #63
