### 2024-07-09 913a697
Update test_r.yaml
Install conda outside r
