### 2024-07-09 80a5a3d
Update test_r.yaml
deactivate (likely fail on mac)
