+++ .github/workflows/test_r.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Conda install
         shell: bash -el {0}
         run: |
-          deactivate test
+          conda deactivate
           conda rename -n test wq_harmonize
           conda install -n wq_harmonize "pytest" "numpy<2.0" "pandas<2.0" "geopandas>=0.10.2, <0.13" "pint>=0.18" "dataretrieval>=1.0, <1.0.5" "requests"
 
