### 2024-07-09 63b91bd
Update test_r.yaml
specify channel
