### 2024-07-09 5b16e46
Update test_r.yaml
Leave test for reticulate to find
