### 2024-07-09 9e29dc6
Update test_r.yaml - miniforge
Use miniforge3, where conda-forge is already default
