### 2024-07-09 ca80c69
Update test_r.yaml
Try read out
