+++ .github/workflows/test_r.yaml
@@ -39,6 +39,12 @@ jobs:
           conda install "pytest" "numpy<2.0" "pandas<2.0" "geopandas>=0.10.2, <0.13" "pint>=0.18" "dataretrieval>=1.0, <1.0.5" "requests"
           conda create --name wq_harmonize --clone test
 
+      - name: Print Conda environment
+        shell: bash -l {0}
+        run: |
+          conda info
+          conda list
+
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
         with:
