### 2024-07-09 4152e90
Update test_r.yaml
try pip installing from path
