### 2024-07-09 82b2861
Update test_r.yaml
re-enable
