### 2024-07-09 a9d8e2c
Update test_r.yaml
pip install then clone
