### 2024-07-09 26116ad
Implementing suggested ruff rules
