+++ harmonize_wq/visualize.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """Functions to help visualize data."""
 from math import sqrt
-import pandas
 import geopandas
+import pandas
 from harmonize_wq.wrangle import merge_tables
 
 
