+++ harmonize_wq/clean.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """Functions to clean/correct additional columns in subset/entire dataset."""
 #from warnings import warn
-from numpy import nan
 import dataretrieval.utils
+from numpy import nan
 from harmonize_wq.convert import convert_unit_series
 from harmonize_wq.domains import accepted_methods
 #from harmonize_wq.wrangle import add_activities_to_df
