+++ harmonize_wq/wq_data.py
@@ -31,7 +31,7 @@ def units_dimension(series_in, units, ureg=None):
     Examples
     --------
     Build series to use as input:
-    
+
     >>> from pandas import Series
     >>> unit_series = Series(['mg/l', 'mg/ml', 'g/kg'])
     >>> unit_series
@@ -41,7 +41,7 @@ def units_dimension(series_in, units, ureg=None):
     dtype: object
 
     Get list of unique units not in desired units dimension 'mg/l':
-    
+
     >>> from harmonize_wq import wq_data
     >>> wq_data.units_dimension(unit_series, units='mg/l')
     ['g/kg']
@@ -85,11 +85,11 @@ class WQCharData():
     units : str
         Units all results in out_col column will be converted into.
         Default units are returned from :func:`domains.OUT_UNITS` [out_col].
-    
+
     Examples
     --------
     Build pandas DataFrame to use as input:
-    
+
     >>> from pandas import DataFrame
     >>> from numpy import nan
     >>> df = DataFrame({'CharacteristicName': ['Phosphorus', 'Temperature, water',],
@@ -100,7 +100,7 @@ class WQCharData():
        CharacteristicName  ResultMeasure/MeasureUnitCode ResultMeasureValue
     0          Phosphorus                            NaN                1.0
     1  Temperature, water                            NaN               10.0
-    
+
     >>> from harmonize_wq import wq_data
     >>> wq = wq_data.WQCharData(df, 'Phosphorus')
     >>> wq.df
@@ -109,7 +109,7 @@ class WQCharData():
     1  Temperature, water                            NaN  ...   NaN         NaN
     <BLANKLINE>
     [2 rows x 5 columns]
-    
+
     >>> wq.df.columns
     Index(['CharacteristicName', 'ResultMeasure/MeasureUnitCode',
            'ResultMeasureValue', 'Units', 'Phosphorus'],
@@ -195,10 +195,10 @@ class WQCharData():
 
     def _unit_qa_flag(self, trouble, flag_col=None):
         """Generate a QA_flag flag string for the units column.
-        
+
         If unit_col is a copy flag_col can specify the original column name for
         the flag. The default units, self.units replaces the problem unit.
-    
+
         Parameters
         ----------
         trouble : str
@@ -206,7 +206,7 @@ class WQCharData():
         flag_col : str, optional
             String to use when referring to the unit_col.
             The default None uses WQCharData.col.unit_out instead.
-    
+
         Returns
         -------
         string
@@ -219,7 +219,7 @@ class WQCharData():
 
     def _replace_in_col(self, col, old_val, new_val, mask=None):
         """Replace string throughout column, filter rows to skip by mask.
-    
+
         Parameters
         ----------
         df_in : pandas.DataFrame
@@ -233,12 +233,12 @@ class WQCharData():
         mask : pandas.Series
             Row conditional mask to only update a sub-set of rows.
             The default None uses 'CharacteristicName' mask instead.
-    
+
         Returns
         -------
         df_in : pandas.DataFrame
             Updated DataFrame.
-    
+
         """
         if mask is None:
             mask = self.c_mask
@@ -254,7 +254,7 @@ class WQCharData():
 
     def _dimension_handling(self, unit, quant=None, ureg=None):
         """Handle and routes common dimension conversions/contexts.
-    
+
         Parameters
         ----------
         unit : str
@@ -264,14 +264,14 @@ class WQCharData():
         ureg : pint.UnitRegistry, optional
             Unit Registry Object with any custom units defined.
             The default is None
-    
+
         Returns
         -------
         dict
             Dictionary with old_unit:new_unit.
         list
             List of Mole (substance) units.
-    
+
         """
         units = self.units
         if ureg is None:
@@ -301,7 +301,7 @@ class WQCharData():
 
[…]
