+++ .github/workflows/test.yml
@@ -3,7 +3,7 @@ name: test
 on:
   push:
     branches:
-      [main, dev]
+      [main, pyOpenSci-review]
   pull_request:
     branches:
       [main]
