+++ harmonize_wq/wrangle.py
@@ -9,7 +9,7 @@ from harmonize_wq.clean import datetime, df_checks, harmonize_depth
 
 def split_table(df_in):
     """Split DataFrame columns axis into main and characteristic based.
-    
+
     Splits :class:`pandas.DataFrame` in two, one with main results columns and
     one with Characteristic based metadata.
 
@@ -32,11 +32,11 @@ def split_table(df_in):
 
     Examples
     --------
-    See any of the 'Simple' notebooks found in 
+    See any of the 'Simple' notebooks found in
     `demos <https://github.com/USEPA/harmonize-wq/tree/main/demos>`_ for
     examples of how this function is used to divide the table into columns of
     interest (main_df) and characteristic specific metadata (chars_df).
-    
+
     """
     # Run datetime on activity fields if not already done
     if 'Activity_datetime' not in list(df_in.columns):
@@ -59,7 +59,7 @@ def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
 
     Values are moved from the result_col in df_in to a new column where the
     column name is col_prefix + characteristic.
-    
+
     Parameters
     ----------
     df_in : pandas.DataFrame
@@ -76,11 +76,11 @@ def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
 
     Examples
     --------
-    See any of the 'Simple' notebooks found in 
+    See any of the 'Simple' notebooks found in
     `demos <https://github.com/USEPA/harmonize-wq/tree/main/demos>`_ for
     examples of how this function is used to split the QA column into multiple
     characteristic specific QA columns.
-    
+
     """
     # TODO: is this function doing too much?
     df_out = df_in.copy()
@@ -139,7 +139,7 @@ def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
 
 def collapse_results(df_in, cols=None):
     """Group rows/results that seems like the same sample.
-    
+
     Default columns are organization, activity, location, and datetime.
 
     Parameters
@@ -156,11 +156,11 @@ def collapse_results(df_in, cols=None):
 
     Examples
     --------
-    See any of the 'Simple' notebooks found in 
+    See any of the 'Simple' notebooks found in
     `demos <https://github.com/USEPA/harmonize-wq/tree/main/demos>`_ for
     examples of how this function is used to combine rows with the same sample
     organization, activity, location, and datetime.
-    
+
     """
     df = df_in.copy()
 
@@ -241,8 +241,8 @@ def collapse_results(df_in, cols=None):
 
 def get_activities_by_loc(characteristic_names, locations):
     """Segment batch what_activities.
-    
-    Warning this is not fully implemented and may not stay. Retrieves in batch 
+
+    Warning this is not fully implemented and may not stay. Retrieves in batch
     using :func:`dataretrieval.what_activities`.
 
     Parameters
@@ -256,7 +256,7 @@ def get_activities_by_loc(characteristic_names, locations):
     -------
     activities : pandas.DataFrame
         Combined activities for locations.
-        
+
     Examples
     --------
     See :func:`wrangle.add_activities_to_df`
@@ -294,26 +294,26 @@ def add_activities_to_df(df_in, mask=None):
     Examples
     --------
     Build example df_in table from harmonize_wq tests to use in place of Water
-    Quality Portal query response, this table has 'Temperature, water' and 
+    Quality Portal query response, this table has 'Temperature, water' and
[…]
