+++ harmonize_wq/__init__.py
@@ -1,5 +1,5 @@
-from harmonize_wq import harmonize
-from importlib.metadata import version, PackageNotFoundError
+from importlib.metadata import PackageNotFoundError, version
+from harmonize_wq import harmonize as harmonize
 
 try:
     __version__ = version('harmonize_wq')
