### 2024-07-09 a8f5ccd
First attempt w/ pre-commit
