+++ .pre-commit-config.yml
@@ -1,0 +1,8 @@
+-   repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.5.0
+    hooks:
+    - id: ruff
+      args: ["--fix", "--show-fixes"]
+      types_or: [ python, jupyter ]
+    - id: ruff-format
+      types_or: [ python, jupyter ]
