+++ .github/workflows/lint.yml
@@ -1,0 +1,27 @@
+name: lint
+
+on:
+  push:
+    branches:
+      [pyOpenSci-review]
+  pull_request:
+    branches:
+      [pyOpenSci-review]
+
+    
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+         
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.x
+      
+      - name: Lint
+	uses: pre-commit/action@v3.0.1
+        with:
+          extra_args: --all-files --show-diff-on-failure
