+++ .github/workflows/lint.yml
@@ -20,8 +20,8 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: 3.x
-      
+
       - name: Lint
-	uses: pre-commit/action@v3.0.1
+	      uses: pre-commit/action@v3.0.1
         with:
           extra_args: --all-files --show-diff-on-failure
