+++ .github/workflows/lint.yml
@@ -21,7 +21,6 @@ jobs:
         with:
           python-version: 3.x
 
-      - name: Lint
-	      uses: pre-commit/action@v3.0.1
+      - uses: pre-commit/action@v3.0.1
         with:
           extra_args: --all-files --show-diff-on-failure
