+++ .pre-commit-config.yaml
@@ -1,3 +1,4 @@
+repos:
 -   repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.5.0
     hooks:
