### 2024-07-11 36e9537
Reduce .pre-commit-config.yaml
Reduce what files it is run on
