+++ .pre-commit-config.yaml
@@ -4,6 +4,4 @@ repos:
     hooks:
     - id: ruff
       args: ["--fix", "--show-fixes"]
-      types_or: [ python, jupyter ]
     - id: ruff-format
-      types_or: [ python, jupyter ]
