### 2024-07-22 #157 @jbousquin
@Batalex If you would like additional links/line numbers just let me know:

_Typos_
**should be resolved as suggested**

_Nitpicks_
_nitpick (general)_
**should be resolved as suggested**

_nitpick (domain.py): there is no need for a raw string for TADA_DATA_URL_
**This url is only used once at the moment, but is currently a raw string (1) to allow it to be easily integrated into feature adds (i.e., intend to use it more places, especially w/ WQX 2->3), and (2) for easier maintenance given the repo is still underdevelopment (e.g., like when the url recently changed).**

_Discussions_
**Kept it in convert module because fewer module references made ensuring no circular references easier. Already importing registry_adds_list from domains so there isn't a strong reason not to move it there if the need arises in the future.**

_Suggestions_
_suggestion (domain.py): In harmonize_TADA_dict, we could use a groupby operation to avoid looping through the dataframe using python. TOCHECK_
**should be resolved as suggested, was there more to the TOCHECK?**

_suggestion (domain.py): We could replace the following pattern for x in list(set(pandas_series)) by using the .unique method_
**should be resolved as suggested**

_suggestion (domain.py, basic.py): out_col_lookup does not need to be a function. Same for all other functions returning a dict. If we make those simple module-level dicts, we can still list the sources in the module docstring._
**These have been updated to be module-level dicts, but I'm not sure on how you are proposing the docstrings could be included. Hate to lose all the examples etc. on these, have you seen this in documentation for other projects you could point me to?**

_suggestion (convert.py): We could add "references" sections in the docstrings so that the sources are present in the website and not only in the source code._
**When a conversion function has equation or methods references the documentation has a reference section for that  (e.g., conductivity_to_PSU). However, if the information is for code/checks then it goes in as a comment in the code (e.g., the url in DO_concentration get to a converter written in JS). In those cases is it adequate/suggested to add contextual comments, e.g., # To check compare against:**

_suggestion (basis.py, general): By using pandas' methods, we could streamline a little some operations. The choice is ultimately yours; I prefer using existing methods over rolling my own implementations, even if that means that other folks need to go to the documentation website to understand what is going on._

**I agree on using existing methods, I really tried to implement this suggestion but ran into issues. In the provided example if there are existing values in columns those need to be preserved. That can be done with an if/else. Additionally, numpy.where will coerce the other values (y) to the dtype which is problematic for nan. Do-able, but more complex than the current solution.**

_Todos_
_pyproject.toml & init_
**should be resolved as suggested**
_basis.py: regroup conditions in update_result_basis_
**Admittedly these additional basis columns haven't received much attention yet (not frequently leveraged by those entering data), and it was coded this way to make it easy to come back to and write additional specific handling. For now we combined weight/time, left particuleSize as is with added notes specific to it's handling.**

_contributing.rst_
**Added dev section**

_Issues_
_domain.py: dependencies_
**Added the suggested dependencies (stop short of pandas but did include numpy). pyproj.toml should populate depends from requirements now - decreasing maintenance/risk of differences.**
_domain.py: specify exception expected by re_case_
**Resolved as suggested**
_Circular dependencies_
**should be resolved as suggested**

_General recommendations_

**To summarize, working on implementing black. All the code changes are sitting on the pyOpenSci-review branch. It runs locally as suggested in your PR. I'm trying to get my head around pre-commits so that contributors will have style/format checks without having to run it locally.**
