### 2024-07-23 12613cf
Update test_r.yaml
Drop test on branch commits
