### 2024-07-23 f4f53b7
62   r test ci (#86)
Update test_r.yaml to install conda outside r, specifically miniforge, then run on env from setup with current package (vs pip installing main)
