### 2024-07-29 414ca16
Update domains.py
Doesn't need to be raw string (see Batalex https://github.com/pyOpenSci/software-submission/issues/157#issuecomment-2256694336)
