### 2024-07-30 a8c1c1a
Dict doc strings as module level attributes
