### 2024-07-30 8725d58
Update to main (#88)
* Update domains.py

'Field' -> 'Field***'

* 62   r test ci (#86)

Update test_r.yaml to install conda outside r, specifically miniforge, then run on env from setup with current package (vs pip installing main)
