### 2024-07-30 dc16881
Update .pre-commit-config.yaml
From issue: pass_filenames: false in the pre-commit config so that the file discovery is done by Ruff taking into account the includes and excludes configured by the user in their pyproject.toml
