+++ .pre-commit-config.yaml
@@ -4,4 +4,5 @@ repos:
     hooks:
     - id: ruff
       args: ["--fix", "--show-fixes"]
+      pass_filenames: false
     - id: ruff-format
