### 2024-07-30 a390231
Update .pre-commit-config.yaml
Try updating to patch version and specify config in args.
