### 2024-07-30 46c0aca
Update pyproject.toml
try withouth 'docstring-code-format = true' as this may override other settings.
