### 2024-08-01 e44f87b
Update pyproject.toml
Try to get pre-commit to see config
