+++ pyproject.toml
@@ -34,7 +34,7 @@ dependencies = { file = ["requirements.txt"] }
 [tool.setuptools.dynamic.optional-dependencies]
 dev = { file = ["requirements-dev.txt"] }
 
-[tool.ruff.lint]
+[tool.ruff]
 select = [
     "E",
     "F",
@@ -42,7 +42,7 @@ select = [
     "I"
 ]
 
-[tool.ruff.lint.isort]
+[tool.ruff.isort]
 known-first-party = ["harmonize_wq"]
 
 [tool.ruff.format]
