### 2024-08-01 2a23d8d
Update pyproject.toml
Warning message, so it is getting these settings from the toml?
