### 2024-08-01 97bbff2
Update basis.py
E501
