### 2024-08-01 20f4214
Update basis.py
Moved constant doc-string to module level
