### 2024-08-01 6ad18cb
Update conf.py
lint/format edits
