+++ docs/source/conf.py
@@ -9,28 +9,29 @@
 # If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 
+import doctest
 import os
 import sys
-import doctest
+
 
 sys.path.insert(0, os.path.abspath(".."))
 sys.path.insert(0, os.path.abspath("../.."))
 
-#from importlib.metadata import version
-#import harmonize_wq
+# from importlib.metadata import version
+# import harmonize_wq
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'harmonize_wq'
-copyright = '2023, US Environmental Protection Agency'
-author = 'Justin Bousquin (US Environmental Protection Agency)'
+project = "harmonize_wq"
+copyright = "2023, US Environmental Protection Agency"
+author = "Justin Bousquin (US Environmental Protection Agency)"
 
 # ToDO:single source version
 version = "0.4.0"
-#release = version(project)
-#release = harmonize_wq.__version__
-#version = '.'.join(release.split('.')[:2])
+# release = version(project)
+# release = harmonize_wq.__version__
+# version = ".".join(release.split(".")[:2])
 
 
 # -- General configuration ---------------------------------------------------
@@ -50,11 +51,11 @@ extensions = [
 
 autosummary_generate = True  # Turn on sphinx.ext.autosummary
 
-# Remove 'view source code' from top of page (for html, not python)
+# Remove "view source code" from top of page (for html, not python)
 html_show_sourcelink = False
 
-templates_path = ['_templates']
-exclude_patterns = ['_build', '_templates']
+templates_path = ["_templates"]
+exclude_patterns = ["_build", "_templates"]
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -65,12 +66,12 @@ html_static_path = []
 
 # Readthedocs theme (may be useful for actions)
 # This line from docs.readthedocs.org, on_rtd is whether on readthedocs.org
-#on_rtd = os.environ.get("READTHEDOCS", None) == "True"
-#if not on_rtd:  # only import and set the theme if we're building docs locally
+# on_rtd = os.environ.get("READTHEDOCS", None) == "True"
+# if not on_rtd:  # only import and set the theme if we're building docs locally
 #    import sphinx_rtd_theme
 #    html_theme = "sphinx_rtd_theme"
 #    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
-#html_css_files = ["readthedocs-custom.css"] # Override some CSS settings
+# html_css_files = ["readthedocs-custom.css"] # Override some CSS settings
 
 
 # -- Options for doctest ------------------------------------------------------
@@ -78,7 +79,7 @@ html_static_path = []
 doctest.testmod(optionflags=doctest.NORMALIZE_WHITESPACE)
 
 # Should enable IGNORE_RESULT option
-IGNORE_RESULT = doctest.register_optionflag('IGNORE_RESULT')
+IGNORE_RESULT = doctest.register_optionflag("IGNORE_RESULT")
 
 OutputChecker = doctest.OutputChecker
 class CustomOutputChecker(OutputChecker):
@@ -101,10 +102,10 @@ napoleon_numpy_docstring = True
 
 # -- Options for sphinx-contrib\apidoc ----------------------------------------
 # NOT currently using apidoc
-#apidoc_separate_modules = True
-#apidoc_module_dir = "../harmonize_wq"
-#apidoc_excluded_paths = ["tests"]
-#apidoc_module_first = True
+# apidoc_separate_modules = True
+# apidoc_module_dir = "../harmonize_wq"
[…]
