### 2024-08-01 c5093d7
Update pyproject.toml
Without single checking if double is default
