### 2024-08-02 e37c33f
Update pyproject.toml
Will move to one or the other (likely default double for ease), but trying to post-pone to work through diff
