+++ pyproject.toml
@@ -44,3 +44,6 @@ select = [
 
 [tool.ruff.lint.isort]
 known-first-party = ["harmonize_wq"]
+
+[tool.ruff.format]
+quote-style = "preserve"
