+++ harmonize_wq/domains.py
@@ -11,7 +11,7 @@ accepted_methods : dict
   Get accepted methods for each characteristic. Dictionary where key is
   characteristic column name and value is list of dictionaries each with Source
   and Method keys.
-  
+
   Notes
   -----
   Source should be in 'ResultAnalyticalMethod/MethodIdentifierContext'
@@ -68,66 +68,66 @@ import pandas
 import requests
 
 
-BASE_URL = 'https://cdx.epa.gov/wqx/download/DomainValues/'
-TADA_DATA_URL = 'https://raw.githubusercontent.com/USEPA/EPATADA/'
+BASE_URL = "https://cdx.epa.gov/wqx/download/DomainValues/"
+TADA_DATA_URL = "https://raw.githubusercontent.com/USEPA/EPATADA/"
 
-UNITS_REPLACE = {'Secchi': {},
-                 'DO': {'%': 'percent'},
-                 'Temperature': {},
-                 'Salinity': {'ppt': 'ppth',
-                              '0/00': 'ppth'},
-                 'pH': {'None': 'dimensionless',
-                        'std units': 'dimensionless'},
-                 'Nitrogen': {'cm3/g @STP': 'cm3/g',
-                              'cm3/g STP': 'cm3/g',
-                              '%': 'percent'},
-                 'Conductivity': {'uS': 'uS/cm',
-                                  'umho': 'umho/cm'},
-                 'Carbon': {'% by wt': '%',
-                            '%': 'percent'},
-                 'Chlorophyll': {'mg/cm3': 'mg/cm**3',
-                                 'mg/m3': 'mg/m**3',
-                                 'mg/m2': 'mg/m**3',
-                                 'ug/cm3': 'ug/cm**3'},
-                 'Turbidity': {'mg/l SiO2': 'SiO2',
-                               'ppm SiO2': 'SiO2'},
-                 'Sediment': {'%': 'percent'},
-                 'Fecal_Coliform': {'#/100ml': 'CFU/(100ml)',
-                                    'CFU': 'CFU/(100ml)',
-                                    'MPN': 'MPN/(100ml)'},
-                 'E_coli': {'#/100ml': 'CFU/(100ml)',
-                            'CFU': 'CFU/(100ml)',
-                            'MPN': 'MPN/(100ml)'},
-                 'Phosphorus': {'%': 'percent'},
-                 }
+UNITS_REPLACE = {
+    "Secchi": {},
+    "DO": {"%": "percent"},
+    "Temperature": {},
+    "Salinity": {"ppt": "ppth", "0/00": "ppth"},
+    "pH": {"None": "dimensionless", "std units": "dimensionless"},
+    "Nitrogen": {"cm3/g @STP": "cm3/g", "cm3/g STP": "cm3/g", "%": "percent"},
+    "Conductivity": {"uS": "uS/cm", "umho": "umho/cm"},
+    "Carbon": {"% by wt": "%", "%": "percent"},
+    "Chlorophyll": {
+        "mg/cm3": "mg/cm**3",
+        "mg/m3": "mg/m**3",
+        "mg/m2": "mg/m**3",
+        "ug/cm3": "ug/cm**3",
+    },
+    "Turbidity": {"mg/l SiO2": "SiO2", "ppm SiO2": "SiO2"},
+    "Sediment": {"%": "percent"},
+    "Fecal_Coliform": {
+        "#/100ml": "CFU/(100ml)",
+        "CFU": "CFU/(100ml)",
+        "MPN": "MPN/(100ml)",
+    },
+    "E_coli": {"#/100ml": "CFU/(100ml)", "CFU": "CFU/(100ml)", "MPN": "MPN/(100ml)"},
[…]
