+++ harmonize_wq/tests/test_harmonize_WQP.py
@@ -26,30 +26,30 @@ from harmonize_wq import visualize as viz
 DIRPATH = os.path.dirname(os.path.realpath(__file__))
 
 # Test datasets
-test_dir = os.path.join(DIRPATH, 'data')
+test_dir = os.path.join(DIRPATH, "data")
 
-AOI_URL = r'https://github.com/USEPA/Coastal_Ecological_Indicators/raw/master/DGGS_Coastal/temperature_data/TampaBay.geojson'
+AOI_URL = r"https://github.com/USEPA/Coastal_Ecological_Indicators/raw/master/DGGS_Coastal/temperature_data/TampaBay.geojson"
 
 # results for dataretrieval.wqp.what_sites(**query)
-STATIONS = pandas.read_csv(os.path.join(test_dir, 'wqp_sites.txt'))
+STATIONS = pandas.read_csv(os.path.join(test_dir, "wqp_sites.txt"))
 # These are split by parameter sets of 2 to keep them small but not mono-param
 # 'Phosphorus' & 'Temperature, water'
-NARROW_RESULTS = pandas.read_csv(os.path.join(test_dir, 'wqp_results.txt'))
-ACTIVITIES = pandas.read_csv(os.path.join(test_dir, 'wqp_activities.txt'))
+NARROW_RESULTS = pandas.read_csv(os.path.join(test_dir, "wqp_results.txt"))
+ACTIVITIES = pandas.read_csv(os.path.join(test_dir, "wqp_activities.txt"))
 # 'Depth, Secchi disk depth' & Dissolved Oxygen
-NARROW_RESULTS1 = pandas.read_csv(os.path.join(test_dir, 'wqp_results1.txt'))
+NARROW_RESULTS1 = pandas.read_csv(os.path.join(test_dir, "wqp_results1.txt"))
 # pH & Salinity
-NARROW_RESULTS2 = pandas.read_csv(os.path.join(test_dir, 'wqp_results2.txt'))
+NARROW_RESULTS2 = pandas.read_csv(os.path.join(test_dir, "wqp_results2.txt"))
 # Nitrogen & Conductivity
-NARROW_RESULTS3 = pandas.read_csv(os.path.join(test_dir, 'wqp_results3.txt'))
+NARROW_RESULTS3 = pandas.read_csv(os.path.join(test_dir, "wqp_results3.txt"))
 # Chlorophyll_a & Organic_carbon
-NARROW_RESULTS4 = pandas.read_csv(os.path.join(test_dir, 'wqp_results4.txt'))
+NARROW_RESULTS4 = pandas.read_csv(os.path.join(test_dir, "wqp_results4.txt"))
 # Turbidity & Sediment
-NARROW_RESULTS5 = pandas.read_csv(os.path.join(test_dir, 'wqp_results5.txt'))
+NARROW_RESULTS5 = pandas.read_csv(os.path.join(test_dir, "wqp_results5.txt"))
 # Nutrients and sediment additional characteristics
 # NARROW_RESULTS6 = pandas.read_csv(os.path.join(test_dir, 'wqp_results6.txt'))
 # Fecal Coliform and Ecoli
-NARROW_RESULTS7 = pandas.read_csv(os.path.join(test_dir, 'wqp_results7.txt'))
+NARROW_RESULTS7 = pandas.read_csv(os.path.join(test_dir, "wqp_results7.txt"))
 
 # fixture to eventually test output writing (.shp)
 # @pytest.fixture(scope="session")
@@ -69,11 +69,13 @@ def test_get_bounding_box():
     AOI : geopandas.GeoDataFrame
         Geodataframe for Tampa Bay read from github
     """
-    expected = ['-82.76095952246396',
-                '27.47487752677648',
-                '-82.37480995151799',
[…]
