+++ harmonize_wq/wq_data.py
@@ -46,7 +46,7 @@ def units_dimension(series_in, units, ureg=None):
     >>> wq_data.units_dimension(unit_series, units='mg/l')
     ['g/kg']
     """
-    #TODO: this should be a method
+    # TODO: this should be a method
     if ureg is None:
         ureg = pint.UnitRegistry()
     dim_list = []  # List for units with mismatched dimensions
@@ -59,7 +59,7 @@ def units_dimension(series_in, units, ureg=None):
     return dim_list
 
 
-class WQCharData():
+class WQCharData:
     """Class for specific characteristic in Water Quality Portal results.
 
     Parameters
@@ -121,16 +121,17 @@ class WQCharData():
         df_out = df_in.copy()
         # self.check_df(df)
         df_checks(df_out)
-        c_mask = df_out['CharacteristicName'] == char_val
+        c_mask = df_out["CharacteristicName"] == char_val
         self.c_mask = c_mask
         # Deal with units: set out = in
-        cols = {'unit_in': 'ResultMeasure/MeasureUnitCode',
-                'unit_out': 'Units',
-                'measure': 'ResultMeasureValue',
-                'basis': 'Speciation', }
+        cols = {
+            "unit_in": "ResultMeasure/MeasureUnitCode",
+            "unit_out": "Units",
+            "measure": "ResultMeasureValue",
+            "basis": "Speciation",
+        }
         self.col = SimpleNamespace(**cols)
-        df_out.loc[c_mask, self.col.unit_out] = df_out.loc[c_mask,
-                                                           self.col.unit_in]
+        df_out.loc[c_mask, self.col.unit_out] = df_out.loc[c_mask, self.col.unit_in]
         self.df = df_out
         # Deal with values: set out_col = in
         self.out_col = domains.out_col_lookup[char_val]
@@ -148,14 +149,13 @@ class WQCharData():
         meas_col = self.col.measure
 
         # Coerce bad measures in series to NaN
-        meas_s = pandas.to_numeric(df_out.loc[c_mask, meas_col],
-                                   errors='coerce')
+        meas_s = pandas.to_numeric(df_out.loc[c_mask, meas_col], errors="coerce")
         # Create a list of the bad measures in the series
         bad_measures = [df_out.iloc[i][meas_col] for i in meas_s[meas_s.isna()].index]
         for bad_meas in pandas.unique(bad_measures):
             # Flag each unique bad measure one measure (not row) at a time
             if pandas.isna(bad_meas):
-                flag = f'{meas_col}: missing (NaN) result'
+                flag = f"{meas_col}: missing (NaN) result"
                 cond = c_mask & (df_out[meas_col].isna())
             else:
                 flag = f'{meas_col}: "{bad_meas}" result cannot be used'
@@ -184,14 +184,13 @@ class WQCharData():
             The default None uses WQCharData.col.unit_out instead.
         """
         # QA flag for missing units
-        flag = self._unit_qa_flag('MISSING', flag_col)
+        flag = self._unit_qa_flag("MISSING", flag_col)
         # Update mask for missing units
         units_mask = self.c_mask & self.df[self.col.unit_out].isna()
         self.df = add_qa_flag(self.df, units_mask, flag)  # Assign flag
         # Update with infered unit
         self.df.loc[units_mask, self.col.unit_out] = self.units
         # Note: .fillna(self.units) is slightly faster but hits datatype issues
-
 
     def _unit_qa_flag(self, trouble, flag_col=None):
         """Generate a QA_flag flag string for the units column.
@@ -213,9 +212,9 @@ class WQCharData():
             Flag to use in QA_flag column.
         """
         if flag_col:
-            return f'{flag_col}: {trouble} UNITS, {self.units} assumed'
+            return f"{flag_col}: {trouble} UNITS, {self.units} assumed"
[…]

[…]
