### 2024-08-02 be5cc9a
W605 - try pulling r str out of test doc-string and instead as a comment. Comment shouldn't cause problems but this one has in the past.
