+++ harmonize_wq/location.py
@@ -5,6 +5,7 @@ import pandas
 from dataretrieval import wqp
 from pyproj import Transformer
 from shapely.geometry import shape
+
 from harmonize_wq.clean import add_qa_flag, check_precision, df_checks
 from harmonize_wq.domains import xy_datum
 from harmonize_wq.wrangle import clip_stations
