+++ harmonize_wq/harmonize.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 """Functions to harmonize data retrieved from EPA's Water Quality Portal."""
 from warnings import warn
+
 from numpy import nan
+
 from harmonize_wq import convert
 from harmonize_wq.domains import OUT_UNITS, UNITS_REPLACE
 from harmonize_wq.visualize import print_report
