+++ docs/source/conf.py
@@ -13,7 +13,6 @@ import doctest
 import os
 import sys
 
-
 sys.path.insert(0, os.path.abspath(".."))
 sys.path.insert(0, os.path.abspath("../.."))
 
