+++ harmonize_wq/clean.py
@@ -3,8 +3,10 @@
 #from warnings import warn
 import dataretrieval.utils
 from numpy import nan
+
 from harmonize_wq.convert import convert_unit_series
 from harmonize_wq.domains import accepted_methods
+
 #from harmonize_wq.wrangle import add_activities_to_df
 
 
