### 2024-08-02 c0f42b0
lint conf file
