+++ harmonize_wq/clean.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 """Functions to clean/correct additional columns in subset/entire dataset."""
-#from warnings import warn
+# from warnings import warn
 import dataretrieval.utils
 from numpy import nan
 
 from harmonize_wq.convert import convert_unit_series
 from harmonize_wq.domains import accepted_methods
 
-#from harmonize_wq.wrangle import add_activities_to_df
+# from harmonize_wq.wrangle import add_activities_to_df
 
 
 def datetime(df_in):
@@ -46,18 +46,20 @@ def datetime(df_in):
     [2 rows x 4 columns]
     """
     # Expected columns
-    date, time, tz = ('ActivityStartDate',
-                      'ActivityStartTime/Time',
-                      'ActivityStartTime/TimeZoneCode')
+    date, time, tz = (
+        "ActivityStartDate",
+        "ActivityStartTime/Time",
+        "ActivityStartTime/TimeZoneCode",
+    )
     df_out = df_in.copy()
     # NOTE: even if date, if time is NA datetime is NaT
     df_out = dataretrieval.utils.format_datetime(df_out, date, time, tz)
-    df_out = df_out.rename(columns={'datetime': 'Activity_datetime'})
+    df_out = df_out.rename(columns={"datetime": "Activity_datetime"})
 
     return df_out
 
 
-def harmonize_depth(df_in, units='meter'):
+def harmonize_depth(df_in, units="meter"):
     """Create 'Depth' column with result depth values in consistent units.
 
     New column combines values from the 'ResultDepthHeightMeasure/MeasureValue' column
@@ -107,16 +109,18 @@ def harmonize_depth(df_in, units='meter'):
     """
     df_out = df_in.copy()
     # Default columns
-    meas_col = 'ResultDepthHeightMeasure/MeasureValue'
-    unit_col = 'ResultDepthHeightMeasure/MeasureUnitCode'
+    meas_col = "ResultDepthHeightMeasure/MeasureValue"
+    unit_col = "ResultDepthHeightMeasure/MeasureUnitCode"
     # Note: there are also 'Activity' cols for both of these & top/bottom depth
 
     df_checks(df_out, [meas_col, unit_col])  # Confirm columns in df
     na_mask = df_out[meas_col].notna()  # Mask NA to speed up processing
     # TODO: if units missing?
-    params = {'quantity_series': df_out.loc[na_mask, meas_col],
-              'unit_series': df_out.loc[na_mask, unit_col],
-              'units': units, }
+    params = {
+        "quantity_series": df_out.loc[na_mask, meas_col],
+        "unit_series": df_out.loc[na_mask, unit_col],
+        "units": units,
+    }
     df_out.loc[na_mask, "Depth"] = convert_unit_series(**params)
 
     # TODO: where result depth is missing use activity depth?
@@ -160,11 +164,13 @@ def df_checks(df_in, columns=None):
     """
     if columns is None:
         # Assign defaults
-        columns = ('ResultMeasure/MeasureUnitCode',
-                   'ResultMeasureValue',
-                   'CharacteristicName')
+        columns = (
+            "ResultMeasure/MeasureUnitCode",
+            "ResultMeasureValue",
+            "CharacteristicName",
+        )
     for col in columns:
-        assert col in df_in.columns, f'{col} not in DataFrame'
+        assert col in df_in.columns, f"{col} not in DataFrame"
 
 
 def check_precision(df_in, col, limit=3):
@@ -191,8 +197,8 @@ def check_precision(df_in, col, limit=3):
     """
     df_out = df_in.copy()
[…]
