+++ harmonize_wq/convert.py
@@ -13,31 +13,32 @@ from numpy import nan
 from harmonize_wq.domains import registry_adds_list
 
 # TODO: does this constant belong here or in domains?
-PERIODIC_MW = {'Organic carbon': 180.16,
-               'C6H12O6': 180.16,
-               'Phosphorus': 30.97,
-               'P': 30.97,
-               'PO4': 94.97,
-               'Nitrogen': 14.01,
-               'N': 14.01,
-               'NO3': 62.01,
-               'NO2': 46.01,
-               'NH4': 18.04,
-               'NH3': 17.03,
-               'SiO3': 76.08,
-               }
+PERIODIC_MW = {
+    "Organic carbon": 180.16,
+    "C6H12O6": 180.16,
+    "Phosphorus": 30.97,
+    "P": 30.97,
+    "PO4": 94.97,
+    "Nitrogen": 14.01,
+    "N": 14.01,
+    "NO3": 62.01,
+    "NO2": 46.01,
+    "NH4": 18.04,
+    "NH3": 17.03,
+    "SiO3": 76.08,
+}
 # Molecular weight assumptions: Organic carbon = C6H12O6
 # NOTE: for a more complete handling of MW: CalebBell/chemicals
 
 u_reg = pint.UnitRegistry()  # For use in wrappers
 # TODO: find more elegant way to do this with all definitions
-for definition in registry_adds_list('Turbidity'):
+for definition in registry_adds_list("Turbidity"):
     u_reg.define(definition)
-for definition in registry_adds_list('Salinity'):
+for definition in registry_adds_list("Salinity"):
     u_reg.define(definition)
 
 
-#timeit: 159.17
+# timeit: 159.17
 # def convert_unit_series(quantity_series, unit_series, units, ureg=None):
 #     # Convert quantities to float if they aren't already (should be)
 #     if quantity_series.dtype=='O':
@@ -53,8 +54,8 @@ for definition in registry_adds_list('Salinity'):
 #     out_list = [val.to(ureg(units)) for val in val_list]
 #     # Re-index to return series
 #     return pandas.Series(out_list, index=quantity_series.index)
-#timeit: 27.08
-def convert_unit_series(quantity_series, unit_series, units, ureg=None, errors='raise'):
+# timeit: 27.08
+def convert_unit_series(quantity_series, unit_series, units, ureg=None, errors="raise"):
     """Convert quantities to consistent units.
 
     Convert list of quantities (quantity_list), each with a specified old unit,
@@ -98,18 +99,18 @@ def convert_unit_series(quantity_series, unit_series, units, ureg=None, errors='
     1    10000.000000000002 milligram / liter
     dtype: object
     """
-    if quantity_series.dtype=='O':
+    if quantity_series.dtype == "O":
         quantity_series = pandas.to_numeric(quantity_series)
     # Initialize classes from pint
     if ureg is None:
         ureg = pint.UnitRegistry()
     Q_ = ureg.Quantity
 
-    lst_series = [pandas.Series(dtype='object')]
+    lst_series = [pandas.Series(dtype="object")]
     # Note: set of series does not preservce order and must be sorted at end
     for unit in list(set(unit_series)):
         # Filter quantity_series by unit_series where == unit
-        f_quant_series = quantity_series.where(unit_series==unit).dropna()
[…]

[…]
