+++ harmonize_wq/convert.py
@@ -3,6 +3,7 @@
 
 Contains several unit conversion functions not in :mod:`pint`.
 """
+
 import math
 from warnings import warn
 
