+++ harmonize_wq/wrangle.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 """Functions to help re-shape the WQP pandas DataFrame."""
+
 import geopandas
 import pandas
 from dataretrieval import wqp
@@ -40,12 +41,12 @@ def split_table(df_in):
 
     """
     # Run datetime on activity fields if not already done
-    if 'Activity_datetime' not in list(df_in.columns):
+    if "Activity_datetime" not in list(df_in.columns):
         df_out = datetime(df_in)
     else:
         df_out = df_in.copy()
     # Run depth if not already done
-    if 'Depth' not in list(df_in.columns):
+    if "Depth" not in list(df_in.columns):
         df_out = harmonize_depth(df_out)
 
     chars_cols = domains.characteristic_cols()  # Characteristic columns list
@@ -55,7 +56,7 @@ def split_table(df_in):
     return main_df, chars_df
 
 
-def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
+def split_col(df_in, result_col="QA_flag", col_prefix="QA"):
     """Move each row value from a column to a characteristic specific column.
 
     Values are moved from the result_col in df_in to a new column where the
@@ -85,21 +86,21 @@ def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
     """
     # TODO: is this function doing too much?
     df_out = df_in.copy()
-    char_list = list(set(df_out['CharacteristicName']))
+    char_list = list(set(df_out["CharacteristicName"]))
 
     # TODO: try/catch on key error
     col_list = [domains.out_col_lookup[char_name] for char_name in char_list]
 
     # TODO: generalize to multi-characteristics other than phosphorus
-    char = 'Phosphorus'
+    char = "Phosphorus"
     if char in char_list:
         i = char_list.index(char)
-        suffix = '_' + domains.out_col_lookup[char]
+        suffix = "_" + domains.out_col_lookup[char]
         col_list[i] = [col for col in df_out.columns if col.endswith(suffix)]
 
     # Drop rows where result na
     for i, char in enumerate(char_list):
-        mask = (df_out['CharacteristicName'] == char)
+        mask = df_out["CharacteristicName"] == char
         if isinstance(col_list[i], list):
             # All columns with that suffix must be nan
             for col in col_list[i]:
@@ -114,12 +115,12 @@ def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
         # Currently written to drop NaN qa flags, to keep them filter on char
         if isinstance(out_col, list):
             for col_out in out_col:
-                new_col = col_prefix + '_' + col_out
+                new_col = col_prefix + "_" + col_out
                 mask = df_out[col_out].notna()
                 df_out.loc[mask, new_col] = df_out.loc[mask, result_col]
         else:
             mask = df_out[out_col].notna()
-            new_col = col_prefix + '_' + out_col
+            new_col = col_prefix + "_" + out_col
             # Create characteristic specific QA field
             df_out.loc[mask, new_col] = df_out.loc[mask, result_col]
 
@@ -130,11 +131,11 @@ def split_col(df_in, result_col='QA_flag', col_prefix='QA'):
 
 
 # def split_unit(series):
-    # If results are being written to another format that does not support
-    # pint objects the units must be recorded. If in the standard ureg it seems
-    # to write them as string, otherwise it errors. Ideally we'd either
-    # transfer the units to within the column name or in a seperate column (not
-    # preffered, only is multiple units).
+# If results are being written to another format that does not support
[…]
