+++ harmonize_wq/visualize.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 """Functions to help visualize data."""
+
 from math import sqrt
 
 import geopandas
@@ -37,26 +38,27 @@ def print_report(results_in, out_col, unit_col_in, threshold=None):
     # Series with just usable results.
     results = results_in[out_col].dropna()
     # Series with infered units
-    inferred = results_in.loc[((results_in[out_col].notna()) &
-                               (results_in[unit_col_in].isna()))]
+    inferred = results_in.loc[
+        ((results_in[out_col].notna()) & (results_in[unit_col_in].isna()))
+    ]
     # Series with just magnitude
     results_s = pandas.Series([x.magnitude for x in results])
     # Number of usable results
-    print(f'-Usable results-\n{results_s.describe()}')
+    print(f"-Usable results-\n{results_s.describe()}")
     # Number measures unused
-    print(f'Unusable results: {len(results_in)-len(results)}')
+    print(f"Unusable results: {len(results_in)-len(results)}")
     # Number of infered result units
-    print(f'Usable results with inferred units: {len(inferred)}')
+    print(f"Usable results with inferred units: {len(inferred)}")
     # Results outside thresholds
     if not threshold:
         # TODO: Default mean +/-1 standard deviation works here but generally 6
-        threshold = {'min': 0.0,
-                     'max': results_s.mean() + (6 * results_s.std())}
-    inside = results_s[(results_s <= threshold['max']) &
-                       (results_s >= threshold['min'])]
+        threshold = {"min": 0.0, "max": results_s.mean() + (6 * results_s.std())}
+    inside = results_s[
+        (results_s <= threshold["max"]) & (results_s >= threshold["min"])
+    ]
     diff = len(results) - len(inside)
     threshold_range = f"{threshold['min']} to {threshold['max']}"
-    print(f'Results outside threshold ({threshold_range}): {diff}')
+    print(f"Results outside threshold ({threshold_range}): {diff}")
 
     # Graphic representation of stats
     inside.hist(bins=int(sqrt(inside.count())))
@@ -126,21 +128,21 @@ def map_counts(df_in, gdf, col=None):
     <Axes: >
     """
     # Column for station
-    loc_id = 'MonitoringLocationIdentifier'
+    loc_id = "MonitoringLocationIdentifier"
     # TODO: col is going to be used to restrict results, if none use all
     if col is not None:
         cols = [loc_id, col]
         df_in = df_in.loc[df_in[col].notna(), cols].copy()
         # TODO: cols needed?
     # Map counts of all results
-    df_cnt = df_in.groupby(loc_id).size().to_frame('cnt')
+    df_cnt = df_in.groupby(loc_id).size().to_frame("cnt")
     df_cnt.reset_index(inplace=True)
 
     # Join it to geometry
-    merge_cols = ['MonitoringLocationIdentifier']
-    gdf_cols = ['geometry', 'QA_flag']
+    merge_cols = ["MonitoringLocationIdentifier"]
+    gdf_cols = ["geometry", "QA_flag"]
     results_df = merge_tables(df_cnt, gdf, gdf_cols, merge_cols)
-    return geopandas.GeoDataFrame(results_df, geometry='geometry')
+    return geopandas.GeoDataFrame(results_df, geometry="geometry")
 
 
 def map_measure(df_in, gdf, col):
@@ -211,7 +213,7 @@ def map_measure(df_in, gdf, col):
     >>> avg_temp.plot(column='mean', cmap='Blues', legend=True)
     <Axes: >
     """
[…]
