### 2024-08-02 9680e35
Add assert for actual2 - where the characteristics specific function is used instead of the generic.
