### 2024-08-02 789b719
Spread out dict read out to reduce line length. White space is already normalized for doc-test so this may pass.
