### 2024-08-02 a504e8f
Spread out building df for wq_dat.WQCharData example.
