### 2024-08-02 6a39fa5
spread out example df for we_date.measure_mask()
