### 2024-08-02 72b78a0
Shotern len of dict for wq_data.replace_unit_str() & wq_data.apply_conversion() examples
