### 2024-08-02 6f57bb4
Last attempt to ignore line too long in docstrings (3)
