+++ harmonize_wq/__init__.py
@@ -3,6 +3,6 @@ from importlib.metadata import PackageNotFoundError, version
 from harmonize_wq import harmonize as harmonize
 
 try:
-    __version__ = version('harmonize_wq')
+    __version__ = version("harmonize_wq")
 except PackageNotFoundError:
     __version__ = "version-unknown"
