### 2024-08-02 5b6a4ed
Update test.yml
Revert back to testing on main only
