### 2024-08-03 878a29f
Py open sci review (#89)
* Implementing suggested ruff rules

* isort

* Fix whitespace (many of these were copied from docs example execution - need to confirm it passes docs tests)

* Run test.yml on push to this branch

* Whitespace

* F401 (redundant alias)

* Missed whitespace

* First attempt w/ pre-commit

* Fix indent

* indent/drop name

* Rename .pre-commit-config.yml to .pre-commit-config.yaml

yAml

* Update .pre-commit-config.yaml

fix file structure

* Reduce .pre-commit-config.yaml

Reduce what files it is run on

* Update domains.py

Doesn't need to be raw string (see Batalex h
[…]
