+++ harmonize_wq/tests/test_harmonize_WQP.py
@@ -1,8 +1,5 @@
 # -*- coding: utf-8 -*-
 """
-This will import when run from CI because sys.path[0] == cur_dir
-DIRPATH = r'L:\Public\jbousqui\Code\GitHub\harmonize-wq\harmonize_wq\tests'
-
 This script doesn't test query/download of the data using dataretrieval,
 instead the script is focused on processing, tidying and harmonizing data
 results from a query read from a ?csv. The exception is the bounding box query
@@ -10,46 +7,45 @@ used to construct the query.
 
 @author: jbousqui
 """
+
 import os
-import pytest
+
 import geopandas
 import pandas
-from harmonize_wq import location
-from harmonize_wq import harmonize
-from harmonize_wq import convert
-from harmonize_wq import wrangle
-from harmonize_wq import clean
-from harmonize_wq import visualize as viz
+import pytest
 
+from harmonize_wq import clean, convert, harmonize, location, wrangle
+from harmonize_wq import visualize as viz
 
 # CI
 DIRPATH = os.path.dirname(os.path.realpath(__file__))
+# DIRPATH = r"L:\Public\jbousqui\Code\GitHub\harmonize-wq\harmonize_wq\tests"
 
 # Test datasets
-test_dir = os.path.join(DIRPATH, 'data')
+test_dir = os.path.join(DIRPATH, "data")
 
-AOI_URL = r'https://github.com/USEPA/Coastal_Ecological_Indicators/raw/master/DGGS_Coastal/temperature_data/TampaBay.geojson'
+AOI_URL = "https://github.com/USEPA/Coastal_Ecological_Indicators/raw/master/DGGS_Coastal/temperature_data/TampaBay.geojson"
 
 # results for dataretrieval.wqp.what_sites(**query)
-STATIONS = pandas.read_csv(os.path.join(test_dir, 'wqp_sites.txt'))
+STATIONS = pandas.read_csv(os.path.join(test_dir, "wqp_sites.txt"))
 # These are split by parameter sets of 2 to keep them small but not mono-param
 # 'Phosphorus' & 'Temperature, water'
-NARROW_RESULTS = pandas.read_csv(os.path.join(test_dir, 'wqp_results.txt'))
-ACTIVITIES = pandas.read_csv(os.path.join(test_dir, 'wqp_activities.txt'))
+NARROW_RESULTS = pandas.read_csv(os.path.join(test_dir, "wqp_results.txt"))
+ACTIVITIES = pandas.read_csv(os.path.join(test_dir, "wqp_activities.txt"))
 # 'Depth, Secchi disk depth' & Dissolved Oxygen
-NARROW_RESULTS1 = pandas.read_csv(os.path.join(test_dir, 'wqp_results1.txt'))
+NARROW_RESULTS1 = pandas.read_csv(os.path.join(test_dir, "wqp_results1.txt"))
 # pH & Salinity
-NARROW_RESULTS2 = pandas.read_csv(os.path.join(test_dir, 'wqp_results2.txt'))
+NARROW_RESULTS2 = pandas.read_csv(os.path.join(test_dir, "wqp_results2.txt"))
 # Nitrogen & Conductivity
-NARROW_RESULTS3 = pandas.read_csv(os.path.join(test_dir, 'wqp_results3.txt'))
+NARROW_RESULTS3 = pandas.read_csv(os.path.join(test_dir, "wqp_results3.txt"))
 # Chlorophyll_a & Organic_carbon
-NARROW_RESULTS4 = pandas.read_csv(os.path.join(test_dir, 'wqp_results4.txt'))
[…]
