+++ harmonize_wq/wq_data.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 """Class for harmonizing data retrieved from EPA's Water Quality Portal."""
+
 from types import SimpleNamespace
 from warnings import warn
+
 import pandas
 import pint
 from numpy import nan
-from harmonize_wq import domains
-from harmonize_wq import basis
-from harmonize_wq.clean import df_checks, add_qa_flag
+
+from harmonize_wq import basis, domains
+from harmonize_wq.clean import add_qa_flag, df_checks
 from harmonize_wq.convert import convert_unit_series, moles_to_mass
 
 
@@ -32,7 +34,7 @@ def units_dimension(series_in, units, ureg=None):
     Examples
     --------
     Build series to use as input:
-    
+
     >>> from pandas import Series
     >>> unit_series = Series(['mg/l', 'mg/ml', 'g/kg'])
     >>> unit_series
@@ -42,12 +44,12 @@ def units_dimension(series_in, units, ureg=None):
     dtype: object
 
     Get list of unique units not in desired units dimension 'mg/l':
-    
+
     >>> from harmonize_wq import wq_data
     >>> wq_data.units_dimension(unit_series, units='mg/l')
     ['g/kg']
     """
-    #TODO: this should be a method
+    # TODO: this should be a method
     if ureg is None:
         ureg = pint.UnitRegistry()
     dim_list = []  # List for units with mismatched dimensions
@@ -60,7 +62,7 @@ def units_dimension(series_in, units, ureg=None):
     return dim_list
 
 
-class WQCharData():
+class WQCharData:
     """Class for specific characteristic in Water Quality Portal results.
 
     Parameters
@@ -86,11 +88,11 @@ class WQCharData():
     units : str
         Units all results in out_col column will be converted into.
         Default units are returned from :func:`domains.OUT_UNITS` [out_col].
-    
+
     Examples
     --------
     Build pandas DataFrame to use as input:
-    
+
     >>> from pandas import DataFrame
     >>> from numpy import nan
     >>> df = DataFrame({'CharacteristicName': ['Phosphorus', 'Temperature, water',],
@@ -101,7 +103,7 @@ class WQCharData():
        CharacteristicName  ResultMeasure/MeasureUnitCode ResultMeasureValue
     0          Phosphorus                            NaN                1.0
     1  Temperature, water                            NaN               10.0
-    
+
     >>> from harmonize_wq import wq_data
     >>> wq = wq_data.WQCharData(df, 'Phosphorus')
     >>> wq.df
@@ -110,7 +112,7 @@ class WQCharData():
     1  Temperature, water                            NaN  ...   NaN         NaN
     <BLANKLINE>
     [2 rows x 5 columns]
-    
+
     >>> wq.df.columns
     Index(['CharacteristicName', 'ResultMeasure/MeasureUnitCode',
            'ResultMeasureValue', 'Units', 'Phosphorus'],
@@ -122,16 +124,17 @@ class WQCharData():
         df_out = df_in.copy()
         # self.check_df(df)
         df_checks(df_out)
-        c_mask = df_out['CharacteristicName'] == char_val
+        c_mask = df_out["CharacteristicName"] == char_val
         self.c_mask = c_mask
         # Deal with units: set out = in
-        cols = {'unit_in': 'ResultMeasure/MeasureUnitCode',
-                'unit_out': 'Units',
-                'measure': 'ResultMeasureValue',
-                'basis': 'Speciation', }
+        cols = {
+            "unit_in": "ResultMeasure/MeasureUnitCode",
+            "unit_out": "Units",
+            "measure": "ResultMeasureValue",
+            "basis": "Speciation",
+        }
[…]
