+++ harmonize_wq/__init__.py
@@ -1,7 +1,8 @@
-from harmonize_wq import harmonize
-from importlib.metadata import version, PackageNotFoundError
+from importlib.metadata import PackageNotFoundError, version
+
+from harmonize_wq import harmonize as harmonize
 
 try:
-    __version__ = version('harmonize_wq')
+    __version__ = version("harmonize_wq")
 except PackageNotFoundError:
     __version__ = "version-unknown"

[…]
