### 2024-08-10 #157 @Batalex
🎉 `harmonize-wq` has been approved by pyOpenSci! Thank you @jbousquin  for submitting `harmonize-wq` and many thanks to @rcaneill and @Jacqui-123 for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/157)`.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉

---
