### 2024-08-13 52ef7f3
Update README.md (#90)
Remove bash prompt from bash commands in README.md
