### 2024-08-15 e57c4f2
Create nb_refresh.yml
Initial commit. Workflow updates demo notebooks
