### 2024-08-15 f13af53
Update nb_refresh.yml
Update when to run
