### 2024-08-15 e8afe8e
Update documentation_test.yml
Remove old branches
