### 2024-08-16 873e9ca
Update nb_refresh.yml
Committer didn't seem to show named bot as commit author, though email listed github-actions[bot]. Docs says to use author name and email (trying w/ default bot name and email this time).
