### 2024-08-18 dfbd3cb
80 refresh demo nb using actions (#91)
* Create nb_refresh.yml
Initial commit. Workflow updates demo notebooks

* Update nb_refresh.yml
Update when to run
Drop on push to this branch
committer not author

* Update documentation_test.yml

Remove old branches

* Update Notebook

* Update nb_refresh.yml

Committer didn't seem to show named bot as commit author, though email listed github-actions[bot]. Docs says to use author name and email (trying w/ default bot name and email this time).

* Update Notebook

---------

Co-authored-by: jbousquin <jbousquin@users.noreply.github.com>
Co-authored-by:
[…]
