### 2024-08-19 9ebe2fe
Reviewed all workflows to bump version to latest and ensure py3.12 is included in testing.
