+++ .github/workflows/test_r.yaml
@@ -1,12 +1,12 @@
 name: R-test
 
 on:
-  push:
-    branches:
-      [main]
   pull_request:
     branches:
       [main]
+  push:
+    branches:
+      [main, dev]
   schedule:
     - cron: '11 1 1 * *'
   # Allows you to run this workflow manually from the Actions tab
@@ -22,7 +22,7 @@ jobs:
     strategy:
       matrix:
         os: [windows-latest, macos-latest]
-        python-version: ['3.8', '3.9', '3.10', '3.11']
+        python-version: ['3.8', '3.9', '3.10', '3.11', '3.12']
         
     steps:
       - uses: actions/checkout@v4
