+++ .github/workflows/test.yml
@@ -1,14 +1,16 @@
 name: test
 
 on:
-  push:
-    branches:
-      [main]
   pull_request:
     branches:
       [main]
+  push:
+    branches:
+      [main, dev]
   schedule:
     - cron: '11 1 1 * *'
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
     
 jobs:
   test:
@@ -19,16 +21,16 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.8', '3.9', '3.10', '3.11']
+        python-version: ['3.8', '3.9', '3.10', '3.11', '3.12']
         #exclude:
         #  - os: windows-latest
         #    python-version: "3.7"
           
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
          
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
       
@@ -62,6 +64,11 @@ jobs:
         run: |
           pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2023.7.16/GDAL-3.7.1-cp311-cp311-win_amd64.whl
           pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2023.7.16/Fiona-1.9.4.post1-cp311-cp311-win_amd64.whl    
+      
+      - if: ${{ (matrix.os == 'windows-latest') && (matrix.python-version == 3.12) }}
+        run: |
+          pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2024.2.18/GDAL-3.8.4-cp312-cp312-win_amd64.whl
+          pip install https://github.com/cgohlke/geospatial-wheels/releases/download/v2024.2.18/fiona-1.9.5-cp312-cp312-win_amd64.whl
       
       - name: Install package and dependencies
         run: |
