+++ .github/workflows/documentation_deploy.yaml
@@ -32,7 +32,7 @@ jobs:
           cd docs
           make html
       - name: Upload artifact
-        uses: actions/upload-pages-artifact@v4
+        uses: actions/upload-pages-artifact@v3
         with:
           # Upload entire repository
           path: './docs/build/html'
