### 2024-08-19 2f4aa79
Update version - some dependencies may have issues with 3.12
