+++ pyproject.toml
@@ -13,7 +13,7 @@ authors = [
 ]
 description = "Package to standardize, clean, and wrangle Water Quality Portal data into more analytic-ready formats"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.8, <3.12"
 keywords = ["USEPA", "water data", "water quality"]
 license = {file = "LICENSE"}
 classifiers = [
