### 2024-08-19 6ccde9d
Drop 3.12 - not supported by pandas<2.0
