### 2024-08-19 4fa7754
Pin to 3.11 since 3.12 is non-compatible with depends
