+++ .github/workflows/nb_refresh.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: 3.x
+          python-version: 3.11
       - name: Update pip and install depends
         run: |
           python -VV
