### 2024-08-21 1a7f5a0
Bump version in pyproject.toml
