### 2024-08-21 427d901
Bump release version in conf.py
