+++ docs/source/conf.py
@@ -27,7 +27,7 @@ copyright = "2023, US Environmental Protection Agency"
 author = "Justin Bousquin (US Environmental Protection Agency)"
 
 # ToDO:single source version
-version = "0.4.0"
+version = "0.5.0"
 # release = version(project)
 # release = harmonize_wq.__version__
 # version = ".".join(release.split(".")[:2])
