### 2024-08-21 e1c5ddb
Add release.yml to publish to PIPY
Publish to PIPY on release using this workflow
