### 2024-08-21 524399a
Update release.yml
quick fix (left out on)
