+++ .github/workflows/release.yml
@@ -2,10 +2,11 @@
 
 name: release
 
-  workflow_dispatch:
-  release:
-    types:
-      - published
+  on:
+    workflow_dispatch:
+    release:
+      types:
+        - published
 
 jobs:
   build_wheels:
