+++ .github/workflows/release.yml
@@ -2,11 +2,11 @@
 
 name: release
 
-  on:
-    workflow_dispatch:
-    release:
-      types:
-        - published
+on:
+  workflow_dispatch:
+  release:
+    types:
+      - published
 
 jobs:
   build_wheels:
