### 2024-08-21 8ab6e0f
Add scheduled and manual runs for nb_refresh.yml
