+++ .github/workflows/nb_refresh.yml
@@ -4,6 +4,10 @@ on:
   pull_request:
     branches:
       [main]
+  schedule:
+    - cron: '15 2 5 * *'
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
 
 jobs:
   notebook_update:
