### 2024-08-21 0ed9749
Update release.yml
Trying to get manual run working
