+++ .github/workflows/release.yml
@@ -3,10 +3,10 @@
 name: release
 
 on:
-  workflow_dispatch:
   release:
     types:
       - published
+  workflow_dispatch:
 
 jobs:
   build_wheels:
