### 2024-08-21 062502a
Update release.yml
Use trusted publisher not token
