+++ .github/workflows/release.yml
@@ -44,7 +44,9 @@ jobs:
   publish-to-pypi:
     needs: [build_wheels, build_sdist]
     runs-on: ubuntu-latest
-    environment: pypi
+    environment:
+      name: pypi
+      url: https://pypi.org/p/harmonize-wq
     permissions:
       id-token: write
     if: github.event_name == 'release' && github.event.action == 'published'
@@ -59,6 +61,3 @@ jobs:
           #path: dist
 
       - uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          user: __token__
-          password: ${{ secrets.PYPI_API_TOKEN }}
