### 2024-08-21 60666ae
Pypi release (#96)
* Add release.yml to publish to PIPY
