### 2024-08-21 6730473
Update release.yml
Attempt to not build/release wheels
