+++ .github/workflows/release.yml
@@ -9,21 +9,21 @@ on:
   workflow_dispatch:
 
 jobs:
-  build_wheels:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-    steps:
-    - uses: actions/checkout@v4
-    - name: Build wheels
-      uses: pypa/cibuildwheel@v2.20.0
-
-    - uses: actions/upload-artifact@v3
-      with:
-        name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
-        path: ./wheelhouse/*.whl
-
+#  build_wheels:
+#    runs-on: ${{ matrix.os }}
+#    strategy:
+#      matrix:
+#        os: [ubuntu-latest, macos-latest, windows-latest]
+#    steps:
+#    - uses: actions/checkout@v4
+#    - name: Build wheels
+#      uses: pypa/cibuildwheel@v2.20.0
+#
+#    - uses: actions/upload-artifact@v3
+#      with:
+#        name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
+#        path: ./wheelhouse/*.whl
+#
   build_sdist:
     name: Build source distribution
     runs-on: ubuntu-latest
@@ -42,7 +42,7 @@ jobs:
           path: dist/*.tar.gz
 
   publish-to-pypi:
-    needs: [build_wheels, build_sdist]
+    needs: build_sdist
     runs-on: ubuntu-latest
     environment:
       name: pypi
@@ -54,9 +54,8 @@ jobs:
       - uses: actions/download-artifact@v4
         with:
           # unpacks all CIBW artifacts into dist/
-          pattern: cibw-*
+          name: cibw-sdist
           path: dist
-          merge-multiple: true
           #name: artifact
           #path: dist
 
