### 2024-08-21 f5603c6
Update release.yml
Build source and wheels
