+++ .github/workflows/release.yml
@@ -25,21 +25,21 @@ jobs:
 #        path: ./wheelhouse/*.whl
 #
   build_sdist:
-    name: Build source distribution
+    name: Build source/wheel distribution
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
         with:
           python-version: '3.11'
-      - name: build the sdist
+      - name: build dist
         run: |
           python -m pip install --upgrade build
-          python -m build --sdist
+          python -m build
       - uses: actions/upload-artifact@v4
         with:
           name: cibw-sdist
-          path: dist/*.tar.gz
+          path: dist
 
   publish-to-pypi:
     needs: build_sdist
