### 2024-08-26 dfef135
Update paper.bib
Added missing refs (Read_2017 & Chegini_2021) from Draft PDF workflow. Added missing DOI for tidy reference from JOSS editorialbot pre-review.
