+++ paper/paper.bib
@@ -33,6 +33,18 @@
   url = {https://doi.org/10.1016/j.envsoft.2021.105210}
 }
 
+@article{Chegini_2021,
+    author = {Chegini, Taher and Li, Hong-Yi and Leung, L. Ruby},
+    doi = {10.21105/joss.03175},
+    journal = {Journal of Open Source Software},
+    month = {10},
+    number = {66},
+    pages = {1--3},
+    title = {{HyRiver: Hydroclimate Data Retriever}},
+    volume = {6},
+    year = {2021}
+}
+
 @Manual{De_Cicco_2022,
   author = {Laura A. {De Cicco} and David Lorenz and Robert M. Hirsch and William Watkins and Mike Johnson},
   title = {dataRetrieval: R packages for discovering and retrieving water data available from U.S. federal hydrologic web services},
@@ -99,6 +111,18 @@
   url = {https://doi.org/10.1002/eap.2130}
 }
 
+@article{Read_2017,
+  author = {Read, E.K., Carr, L., De Cicco, L., Dugan, H.A., Hanson, P.C., Hart, J.A., Kreft, J., Read, J.S. and Winslow, L.A.},
+  year = {2017},
+  title = {Water quality data for national‐scale aquatic research: The Water Quality Portal. Water Resources Research},
+  journal = {Water Resources Research},
+  volume = {53},
+  issue = {2},
+  pages = {1735-1745},
+  doi = {10.1002/2016WR019993},
+  url = {https://doi.org/10.1002/2016WR019993}
+}
+
 @article{Ross_2019,
   author = {Ross, M.R., Topp, S.N., Appling, A.P., Yang, X., Kuhn, C., Butman, D., Simard, M. and Pavelsky, T.M.},
   year = {2019},
@@ -152,7 +176,9 @@
   journal = {The Journal of Statistical Software},
   volume = {59},
   issue = {10},
-  pages = {252–261}
+  pages = {252–261},
+  doi = {10.18637/jss.v059.i10},
+  url = {https://doi.org/10.18637/jss.v059.i10}
 }
 
 @Manual{USEPA_2018,
