### 2024-08-26 1db713c
Update paper.bib
Replace '-' so it gets encoded correctly
