+++ paper/paper.bib
@@ -18,7 +18,7 @@
   volume = {47},
   number = {5},
   pages = {1136-1150},
-  doi = {10.1111/j.1752‐1688.2011.00573.x},
+  doi = {10.1111/j.1752-1688.2011.00573.x},
   url = {https://doi.org/10.1111/j.1752-1688.2011.00573.x}
 }
 
