+++ paper/paper.md
@@ -22,11 +22,11 @@ bibliography: paper.bib
 
 # Summary
 
-The US EPA’s Water Quality Exchange (WQX) allows state environmental agencies, the EPA, other federal agencies, universities, private citizens, and other organizations to provide water quality, biological, and physical data [@Read_2017]. The Water Quality Portal (WQP) is a data warehouse that facilitates access to data stored in large water quality databases, including WQX, in a common format. WQP has become an essential resource with tools to facilitate both data publishing [@USEPA_2018] [@USEPA_2020] and data retrieval [@De_Cicco_2022] [@Hodson_2023]. However, given the variety of data originators and methods, using the data in analysis often requires cleaning to ensure it meets required quality standards and wrangling to get it in a more analytic-ready format. Although there are many examples where this data cleaning or wrangling has been performed [@Bousquin_2021] [@Evans_2021] [@Manning_2020] [@Ross_2019] [@Shen_2020], standardized tools to perform this task will make it less time-intensive, more standardized, and more reproducible. More standardized data cleansing and wrangling allows easier integration of outputs into other tools in the water quality data pipeline, e.g., for integration into hydrologic analysis [@Chegini_2021], dashboards for visualization [@Beck_2021) or decision support tools [@Booth_2011].
+The US EPA’s Water Quality Exchange (WQX) allows state environmental agencies, the EPA, other federal agencies, universities, private citizens, and other organizations to provide water quality, biological, and physical data [@Read_2017]. The Water Quality Portal (WQP) is a data warehouse that facilitates access to data stored in large water quality databases, including WQX, in a common format. WQP has become an essential resource with tools to facilitate both data publishing [@USEPA_2018; @USEPA_2020] and data retrieval [@De_Cicco_2022; @Hodson_2023]. However, given the variety of data originators and methods, using the data in analysis often requires cleaning to ensure it meets required quality standards and wrangling to get it in a more analytic-ready format. Although there are many examples where this data cleaning or wrangling has been performed [@Bousquin_2021;  @Evans_2021; @Manning_2020; @Ross_2019; @Shen_2020], standardized tools to perform this task will make it less time-intensive, more standardized, and more reproducible. More standardized data cleansing and wrangling allows easier integration of outputs into other tools in the water quality data pipeline, e.g., for integration into hydrologic analysis [@Chegini_2021], dashboards for visualization [@Beck_2021] or decision support tools [@Booth_2011].
 
 # Statement of need
 
-Due to the diversity of data originators metadata quality varies and can pose significant challenges preventing WQP from being used as an analysis-ready data set [@Sprague_2017] [@Shaughnessy_2019]. Recognizing the definition of analysis-ready varies depending on the analysis, our goal with harmonize-wq is to provide a robust, flexible, water quality specific framework that will help the data analyst identify differences in data units (including speciation and basis), identify differences in sampling or analytic methods, and resolve data errors using transparent assumptions. Domain experts must decide what data meets their quality standards for data comparability and any thresholds for acceptance or rejection.
+Due to the diversity of data originators metadata quality varies and can pose significant challenges preventing WQP from being used as an analysis-ready data set [@Sprague_2017; @Shaughnessy_2019]. Recognizing the definition of analysis-ready varies depending on the analysis, our goal with harmonize-wq is to provide a robust, flexible, water quality specific framework that will help the data analyst identify differences in data units (including speciation and basis), identify differences in sampling or analytic methods, and resolve data errors using transparent assumptions. Domain experts must decide what data meets their quality standards for data comparability and any thresholds for acceptance or rejection.
 
 # Current Functionality
 
