+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{Beck_2021,
   author = {Beck, M.W., Schrandt, M.N., Wessel, M.R., Sherwood, E.T., Raulerson, G.E., Prasad A.A.B., and Best, B.D.},
   year = {2021},
-  title = {tbeptools: An R package for synthesizing estuarine data for environmental research},
+  title = {{tbeptools}: An {R} package for synthesizing estuarine data for environmental research},
   journal = {Journal of Open Source Software},
   volume = {6},
   number = {65},
