### 2024-09-17 5ab902e
Lock title capitalization
