+++ paper/paper.bib
@@ -13,7 +13,7 @@
 @article{Booth_2011,
   author = {Booth, N.L., Everman, E.J., Kuo, I.L., Sprague, L. and Murphy, L.},
   year = {2011},
-  title = {A Web‐Based Decision Support System for Assessing Regional Water‐Quality Conditions and Management Actions},
+  title = {{A Web‐Based Decision Support System for Assessing Regional Water‐Quality Conditions and Management Actions}},
   journal = {Journal of the American Water Resources Association},
   volume = {47},
   number = {5},
