### 2024-09-17 f10348b
Check/lock title caps throughout
