+++ paper/paper.bib
@@ -25,7 +25,7 @@
 @article{Bousquin_2021,
   author = {Bousquin, J.},
   year = {2021},
-  title = {Discrete Global Grid Systems as scalable geospatial frameworks for characterizing coastal environments},
+  title = {{Discrete Global Grid Systems} as scalable geospatial frameworks for characterizing coastal environments},
   journal = {Environmental Modelling & Software},
   volume = {146},
   pages = {105210},
@@ -47,7 +47,7 @@
 
 @Manual{De_Cicco_2022,
   author = {Laura A. {De Cicco} and David Lorenz and Robert M. Hirsch and William Watkins and Mike Johnson},
-  title = {dataRetrieval: R packages for discovering and retrieving water data available from U.S. federal hydrologic web services},
+  title = {{dataRetrieval: R packages for discovering and retrieving water data available from U.S. federal hydrologic web services}},
   publisher = {U.S. Geological Survey},
   address = {Reston, VA},
   version = {2.7.12},
@@ -71,7 +71,7 @@
 
 @software{Grecco_2021,
   author = {Grecco, H., and Chéron, J.},
-  title = {Pint: Operate and manipulate physical quantities in Python},
+  title = {{Pint: Operate and manipulate physical quantities in Python}},
   version = {1.9},
   year = {2021},
   url = {https://github.com/hgrecco/pint},
@@ -79,7 +79,7 @@
 
 @software{Hodson_2023,
   author = {Hodson, T.O., Hariharan, J.A., Black, S., and Horsburgh, J.S.},
-  title = {dataretrieval (Python): a Python package for discovering and retrieving water data available from U.S. federal hydrologic web services},
+  title = {{dataretrieval (Python): a Python package for discovering and retrieving water data available from U.S. federal hydrologic web services}},
   publisher = {U.S. Geological Survey},
   version = {1.0.2},
   institution = {U.S. Geological Survey},
@@ -90,7 +90,7 @@
 
 @software{Jordahl_2021,
   author = {Kelsey Jordahl, Joris Van den Bossche, Martin Fleischmann, James McBride, Jacob Wasserman, Adrian Garcia Badaracco, Jeffrey Gerard, Alan D. Snow, Jeff Tratner, Matthew Perry, Carson Farmer, Geir Arne Hjelle, Micah Cochran, Sean Gillies, Lucas Culbertson, Matt Bartos, Brendan Ward, Giacomo Caria, Mike Taves, Nick Eubank, sangarshanan, John Flavin, Matt Richards, Sergio Rey, maxalbert, Aleksey Bilogur, Christopher Ren, Dani Arribas-Bel, Daniel Mesejo-León, and Leah Wasser},
-  title = {geopandas/geopandas: v0.10.2},
+  title = {{geopandas/geopandas: v0.10.2}},
   month = oct,
   year = 2021,
   publisher = {Zenodo},
@@ -102,7 +102,7 @@
 @article{Manning_2020,
   author = {Manning, D.W., Rosemond, A.D., Benstead, J.P., Bumpers, P.M. and Kominoski, J.S.},
   year = {2020},
-  title = {Transport of N and P in US streams and rivers differs with land use and between dissolved and particulate forms},
+  title = {{Transport of N and P in US streams and rivers differs with land use and between dissolved and particulate forms}},
   journal = {Ecological Applications},
   volume = {30},
   issue = {6},
@@ -114,7 +114,7 @@
 @article{Read_2017,
   author = {Read, E.K., Carr, L., De Cicco, L., Dugan, H.A., Hanson, P.C., Hart, J.A., Kreft, J., Read, J.S. and Winslow, L.A.},
   year = {2017},
-  title = {Water quality data for national‐scale aquatic research: The Water Quality Portal. Water Resources Research},
[…]
