### 2024-11-01 e4b919a
Update Paper joss (#105)
* Create draft-pdf.yml

Test paper.md compiles still

* Update draft-pdf.yml

In path paper/

* Update draft-pdf.yml

pdf written to same path?

* Update paper.bib

Added missing refs (Read_2017 & Chegini_2021) from Draft PDF workflow. Added missing DOI for tidy reference from JOSS editorialbot pre-review.

* Update paper.bib

Replace '-' so it gets encoded correctly

* Multi-reference formating

* Keep ref capitalization

* Lock title capitalization

* Check/lock title caps throughout

* Minor copy-edit/readability edits.

* Delete .github/workflows/draft-pdf.yml

Thi
[…]
