+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{Beck_2021,
   author = {Beck, M.W., Schrandt, M.N., Wessel, M.R., Sherwood, E.T., Raulerson, G.E., Prasad A.A.B., and Best, B.D.},
   year = {2021},
-  title = {tbeptools: An R package for synthesizing estuarine data for environmental research},
+  title = {{tbeptools}: An {R} package for synthesizing estuarine data for environmental research},
   journal = {Journal of Open Source Software},
   volume = {6},
   number = {65},
@@ -13,19 +13,19 @@
 @article{Booth_2011,
   author = {Booth, N.L., Everman, E.J., Kuo, I.L., Sprague, L. and Murphy, L.},
   year = {2011},
-  title = {A Web‐Based Decision Support System for Assessing Regional Water‐Quality Conditions and Management Actions},
+  title = {{A Web‐Based Decision Support System for Assessing Regional Water‐Quality Conditions and Management Actions}},
   journal = {Journal of the American Water Resources Association},
   volume = {47},
   number = {5},
   pages = {1136-1150},
-  doi = {10.1111/j.1752‐1688.2011.00573.x},
+  doi = {10.1111/j.1752-1688.2011.00573.x},
   url = {https://doi.org/10.1111/j.1752-1688.2011.00573.x}
 }
 
 @article{Bousquin_2021,
   author = {Bousquin, J.},
   year = {2021},
-  title = {Discrete Global Grid Systems as scalable geospatial frameworks for characterizing coastal environments},
+  title = {{Discrete Global Grid Systems} as scalable geospatial frameworks for characterizing coastal environments},
   journal = {Environmental Modelling & Software},
   volume = {146},
   pages = {105210},
@@ -33,9 +33,21 @@
   url = {https://doi.org/10.1016/j.envsoft.2021.105210}
 }
 
+@article{Chegini_2021,
+    author = {Chegini, Taher and Li, Hong-Yi and Leung, L. Ruby},
+    doi = {10.21105/joss.03175},
+    journal = {Journal of Open Source Software},
+    month = {10},
+    number = {66},
+    pages = {1--3},
+    title = {{HyRiver: Hydroclimate Data Retriever}},
+    volume = {6},
+    year = {2021}
+}
+
 @Manual{De_Cicco_2022,
   author = {Laura A. {De Cicco} and David Lorenz and Robert M. Hirsch and William Watkins and Mike Johnson},
-  title = {dataRetrieval: R packages for discovering and retrieving water data available from U.S. federal hydrologic web services},
+  title = {{dataRetrieval: R packages for discovering and retrieving water data available from U.S. federal hydrologic web services}},
   publisher = {U.S. Geological Survey},
   address = {Reston, VA},
   version = {2.7.12},
@@ -59,7 +71,7 @@
 
 @software{Grecco_2021,
   author = {Grecco, H., and Chéron, J.},
-  title = {Pint: Operate and manipulate physical quantities in Python},
+  title = {{Pint: Operate and manipulate physical quantities in Python}},
   version = {1.9},
   year = {2021},
   url = {https://github.com/hgrecco/pint},
@@ -67,7 +79,7 @@
 
 @software{Hodson_2023,
   author = {Hodson, T.O., Hariharan, J.A., Black, S., and Horsburgh, J.S.},
-  title = {dataretrieval (Python): a Python package for discovering and retrieving water data available from U.S. federal hydrologic web services},
[…]
