### 2024-11-01 f58f2f2
Init yml for github-actions & pip (#106)
