### 2024-11-01 0f45249
Update dataretrieval requirement from <1.0.5,>=1.0 to >=1.0,<1.0.12
Updates the requirements on [dataretrieval](https://github.com/DOI-USGS/dataretrieval-python) to permit the latest version.
- [Release notes](https://github.com/DOI-USGS/dataretrieval-python/releases)
- [Commits](https://github.com/DOI-USGS/dataretrieval-python/compare/v1.0.0...v1.0.11)

---
updated-dependencies:
- dependency-name: dataretrieval
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
