### 2024-11-01 8c9ee1e
Update geopandas requirement from <0.13,>=0.10.2 to >=0.10.2,<1.1
Updates the requirements on [geopandas](https://github.com/geopandas/geopandas) to permit the latest version.
- [Release notes](https://github.com/geopandas/geopandas/releases)
- [Changelog](https://github.com/geopandas/geopandas/blob/main/CHANGELOG.md)
- [Commits](https://github.com/geopandas/geopandas/compare/v0.10.2...v1.0.1)

---
updated-dependencies:
- dependency-name: geopandas
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
