+++ requirements.txt
@@ -1,6 +1,6 @@
 numpy<2.0
 pandas<2.0
-geopandas>=0.10.2, <0.13           # Has additional dependencies (pandas, numpy, shapely, fiona, etc.), tested on 0.10.2
+geopandas>=0.10.2, <1.1           # Has additional dependencies (pandas, numpy, shapely, fiona, etc.), tested on 0.10.2
 pint>=0.18          # May work on earlier versions but this is what it was tested on
 dataretrieval>=1.0, <1.0.5  # 1.0.5 will break datetime handling, <1.0 doesn't have required functions
 descartes>=1.1.0           # May be required for mapping in some jupyter notebooks
