### 2024-11-22 #157 @lwasser
hey team. has this package been accepted by JOSS? it looks like we might be able to update the labels, fill out the header and close it if that is the case. please let me know however.
