### 2024-11-22 #157 @jbousquin
Yes - it cleared JOSS and citation info on the repo are updated. Let me know if you need any info/action from me. Thanks!
