### 2025-02-06 #157 @lwasser
fantastic. I am going to close this. I"m sorry for the long delay in response. @jbousquin if you or anyone involved in this review (other maintianers, reviewers etc) would like to join our slack you are welcome to do so. please email me at leah at pyopensci.org and i'll invite you! or share an email here.

if you'd like to share you package with the community, we also welcome you to create a blog post. please just let me know. in the meantime i'll close this issue but we can still chat here if you have questions or need anything else!!

thank you everyone!
